refactor(parser): move blocks/functions/resolve out of parser/mod.rs (part 2/2)
Closes #2263.

Splits the remaining ~7,000 lines of `carina-core/src/parser/mod.rs` into
topical submodules per the issue's target layout:

- `ast.rs`           — all AST structs/enums and `ParsedFile` impl
- `error.rs`         — `ParseError`, `ParseWarning`, error helpers
- `context.rs`       — `ParseContext`, `next_pair`, `first_inner`, `extract_key_string`
- `entry.rs`         — `parse`, `parse_and_resolve`, `parse_expression(_eval)`
- `blocks/provider.rs`     — `parse_provider_block`, `parse_require_statement`
- `blocks/backend.rs`      — `parse_backend_block`, `parse_upstream_state_expr`
- `blocks/resource.rs`     — anonymous + bound + read resources, `parse_block_contents`
- `blocks/module_call.rs`  — `parse_module_call`
- `blocks/attributes.rs`   — `parse_arguments_block`, `parse_attributes_block`,
                             `parse_exports_block`, `extract_lifecycle_config`
- `blocks/state.rs`        — `parse_import_state_block`, `parse_removed_block`,
                             `parse_moved_block`
- `blocks/use_stmt.rs`     — `parse_use_expr`
- `types.rs`         — `parse_type_expr`
- `functions.rs`     — user-defined function parsing/evaluation, `validate_custom_type`
- `let_binding.rs`   — `parse_let_binding_extended` and pipe/compose paths
                       that may produce closures
- `resolve.rs`       — `resolve_resource_refs(_with_config)`,
                       `collect_known_bindings_merged`, `check_identifier_scope`
- `static_eval.rs`   — `is_static_value`, `is_static_eval`, `evaluate_static_value`
- `util.rs`          — `pascal_to_snake`, `snake_to_pascal`, value/eval type names,
                       resource-address parsing, string-literal helpers

`mod.rs` is now declarations and re-exports plus the unchanged inline test
module. Public API (`parse`, `parse_and_resolve`,
`resolve_resource_refs(_with_config)`, `collect_known_bindings_merged`,
`check_identifier_scope`, `validate_custom_type`, `pascal_to_snake`,
`snake_to_pascal`, all AST types, `ForBinding`, `CompareOp`) is unchanged.
