+++ carina-core/src/parser/mod.rs
@@ -2,3765 +2,62 @@
 //!
 //! Convert DSL to AST using pest
 
+mod ast;
+mod blocks;
 mod config;
+mod context;
+mod entry;
+mod error;
 mod expressions;
+mod functions;
+mod let_binding;
+mod resolve;
+mod static_eval;
+mod types;
+mod util;
 
+pub use ast::{
+    ArgumentParameter, AttributeParameter, BackendConfig, DEFERRED_UPSTREAM_PLACEHOLDER,
+    DeferredForExpression, ExportParameter, FnParam, ModuleCall, ParsedFile, ProviderConfig,
+    RequireBlock, ResourceContext, ResourceTypePath, StateBlock, TypeExpr, UpstreamState,
+    UseStatement, UserFunction, UserFunctionBody, ValidateExpr, ValidationBlock,
+};
 pub use config::{DecryptorFn, ProviderContext, ValidatorFn};
+pub use entry::{parse, parse_and_resolve};
+pub(crate) use entry::{parse_expression, parse_expression_eval};
+pub use error::{ParseError, ParseWarning};
+pub(crate) use functions::evaluate_user_function;
+pub use functions::validate_custom_type;
+pub use resolve::{
+    check_identifier_scope, collect_known_bindings_merged, resolve_resource_refs,
+    resolve_resource_refs_with_config,
+};
+pub(crate) use util::{eval_type_name, value_type_name};
+pub use util::{pascal_to_snake, snake_to_pascal};
+
+pub(crate) use ast::{DEFERRED_UPSTREAM_INDEX_PLACEHOLDER, DEFERRED_UPSTREAM_KEY_PLACEHOLDER};
+pub(crate) use blocks::module_call::parse_module_call;
+pub(crate) use blocks::resource::{
+    parse_block_contents, parse_read_resource_expr, parse_resource_expr,
+};
+pub(crate) use context::{ParseContext, extract_key_string, first_inner, next_pair};
 pub use expressions::for_expr::ForBinding;
 pub use expressions::validate_expr::CompareOp;
+pub(crate) use let_binding::LetBindingRhs;
+pub(crate) use static_eval::{evaluate_static_value, is_static_eval, is_static_value};
 
-use expressions::for_expr::{extract_for_iterable_name, parse_for_expr};
-use expressions::if_expr::parse_if_expr;
-use expressions::pipe::parse_coalesce_expr;
-use expressions::primary::parse_primary_eval;
-use expressions::string_literal::{
-    parse_string_literal, parse_string_value, unescape_single_quoted,
-};
-use expressions::validate_expr::parse_validate_expr;
-
-use indexmap::IndexMap;
-
-use crate::eval_value::EvalValue;
-use crate::resource::{Expr, LifecycleConfig, Resource, ResourceId, ResourceKind, Value};
-use crate::schema::{
-    validate_ipv4_address, validate_ipv4_cidr, validate_ipv6_address, validate_ipv6_cidr,
-};
-use crate::version_constraint::VersionConstraint;
-use pest::Parser;
 use pest_derive::Parser;
-use std::collections::{BTreeSet, HashMap, HashSet};
 
 #[derive(Parser)]
 #[grammar = "parser/carina.pest"]
-struct CarinaParser;
-
-/// Parse error
-#[derive(Debug, thiserror::Error)]
-pub enum ParseError {
-    #[error("Syntax error: {0}")]
-    Syntax(#[from] pest::error::Error<Rule>),
-
-    #[error("Invalid expression at line {line}: {message}")]
-    InvalidExpression { line: usize, message: String },
-
-    #[error("Undefined variable: {0}")]
-    UndefinedVariable(String),
-
-    #[error("Invalid resource type: {0}")]
-    InvalidResourceType(String),
-
-    #[error("Duplicate module definition: {0}")]
-    DuplicateModule(String),
-
-    #[error("Duplicate binding at line {line}: {name}")]
-    DuplicateBinding { name: String, line: usize },
-
-    #[error("{}", format_undefined_identifier(name, suggestion.as_deref(), in_scope))]
-    UndefinedIdentifier {
-        name: String,
-        line: usize,
-        /// Edit-distance close match among in-scope bindings, if any.
-        /// Filled by the check site when a `suggest_similar_name` result is
-        /// available; None for hand-constructed errors.
-        suggestion: Option<String>,
-        /// Concrete binding names in scope at the check site, sorted for
-        /// deterministic rendering. Empty when no bindings have been
-        /// declared at all.
-        in_scope: Vec<String>,
-    },
-
-    #[error("Module not found: {0}")]
-    ModuleNotFound(String),
-
-    #[error("Internal parser error: expected {expected} in {context}")]
-    InternalError { expected: String, context: String },
-
-    #[error("Recursive function call detected: {0}")]
-    RecursiveFunction(String),
-
-    #[error("User-defined function error: {0}")]
-    UserFunctionError(String),
-}
-
-/// Render the `UndefinedIdentifier` message. When a close match exists
-/// (`suggestion`), lead with `Did you mean ...?`. Otherwise list the
-/// concrete in-scope names so the reader learns what is available,
-/// followed by the abstract list of binding kinds as a trailing aside.
-/// See #2038.
-fn format_undefined_identifier(
-    name: &str,
-    suggestion: Option<&str>,
-    in_scope: &[String],
-) -> String {
-    if let Some(s) = suggestion {
-        return format!("Undefined identifier `{}`. Did you mean `{}`?", name, s);
-    }
-    let kinds = "let / upstream_state / read / module / function / for / fn / arguments";
-    if in_scope.is_empty() {
-        format!(
-            "Undefined identifier `{}`: no bindings are in scope ({})",
-            name, kinds,
-        )
-    } else {
-        format!(
-            "Undefined identifier `{}`. In-scope names: {} ({})",
-            name,
-            in_scope.join(", "),
-            kinds,
-        )
-    }
-}
-
-/// A structured warning emitted during parsing (non-fatal).
-#[derive(Debug, Clone, PartialEq)]
-pub struct ParseWarning {
-    /// Full source path of the file this warning originated from
-    /// (stamped by `config_loader` after parsing). `None` at parse time;
-    /// always `Some` once the warning reaches CLI/LSP callers.
-    pub file: Option<String>,
-    pub line: usize,
-    pub message: String,
-}
-
-/// Placeholder text for values that depend on an upstream apply.
-pub const DEFERRED_UPSTREAM_PLACEHOLDER: &str = "(known after upstream apply)";
-/// Placeholder reserved for map-binding key variables. Distinct from the
-/// value-var placeholder so expansion can substitute each correctly.
-pub(crate) const DEFERRED_UPSTREAM_KEY_PLACEHOLDER: &str = "(known after upstream apply: key)";
-/// Placeholder reserved for indexed-binding index variables — distinct from
-/// key and value placeholders so `for (i, x) in list` expansion can
-/// substitute `i` with the integer index.
-pub(crate) const DEFERRED_UPSTREAM_INDEX_PLACEHOLDER: &str = "(known after upstream apply: index)";
-
-/// A for-expression whose iterable is unresolved (e.g., upstream_state not yet available).
-/// Captures the structural shape of the loop body so the plan can show what
-/// resources *would* be created once the iterable becomes available.
-/// Also stores enough information to expand the loop later when the iterable
-/// is loaded from upstream_state.
-#[derive(Debug, Clone)]
-pub struct DeferredForExpression {
-    /// Full source path of the file this deferred expression originated
-    /// from (stamped by `config_loader` after parsing).
-    pub file: Option<String>,
-    /// Source line number of the `for` keyword.
-    pub line: usize,
-    /// The for-expression header, e.g., `for account_id in orgs.accounts`.
-    pub header: String,
-    /// The provider-qualified resource type the loop body would produce (e.g., `awscc.sso.Assignment`).
-    pub resource_type: String,
-    /// Attribute template: key → value (concrete values are resolved;
-    /// loop-bound variables remain as `ResourceRef` or placeholder strings).
-    pub attributes: Vec<(String, Value)>,
-    /// The binding address prefix for generated resources (e.g., `_for0`).
-    pub binding_name: String,
-    /// The iterable access path segments (e.g., `["orgs", "accounts"]`).
-    pub iterable_binding: String,
-    pub iterable_attr: String,
-    /// Binding pattern — records the kind (Simple/Indexed/Map) so the expansion
-    /// can verify the resolved iterable's shape matches and substitute the
-    /// correct variable(s).
-    pub binding: ForBinding,
-    /// Template resource for expansion (the for body parsed with placeholders).
-    pub template_resource: Resource,
-}
-
-/// Origin of a resource yielded by [`ParsedFile::iter_all_resources`].
-///
-/// `Direct` means the resource was declared at top-level and its iterable
-/// (if any) resolved at parse time. `Deferred` means the resource is the
-/// template body of a `for` expression whose iterable resolves later;
-/// consumers that care about loop-variable placeholders need the
-/// `DeferredForExpression` reference to filter them out.
-#[derive(Debug, Clone, Copy)]
-pub enum ResourceContext<'a> {
-    Direct,
-    Deferred(&'a DeferredForExpression),
-}
-
-/// Resource type path for typed references (e.g., aws.vpc, aws.security_group)
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct ResourceTypePath {
-    /// Provider name (e.g., "aws")
-    pub provider: String,
-    /// Resource type (e.g., "vpc", "security_group")
-    pub resource_type: String,
-}
-
-impl ResourceTypePath {
-    pub fn new(provider: impl Into<String>, resource_type: impl Into<String>) -> Self {
-        Self {
-            provider: provider.into(),
-            resource_type: resource_type.into(),
-        }
-    }
-
-    /// Parse from a dot-separated string (e.g., "aws.vpc" or "aws.security_group")
-    pub fn parse(s: &str) -> Option<Self> {
-        let parts: Vec<&str> = s.split('.').collect();
-        if parts.len() >= 2 {
-            Some(Self {
-                provider: parts[0].to_string(),
-                resource_type: parts[1..].join("."),
-            })
-        } else {
-            None
-        }
-    }
-}
-
-impl std::fmt::Display for ResourceTypePath {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}.{}", self.provider, self.resource_type)
-    }
-}
-
-/// Type expression for arguments/attributes parameters
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum TypeExpr {
-    String,
-    Bool,
-    Int,
-    Float,
-    /// Schema type identified by name (e.g., "ipv4_cidr", "ipv4_address", "arn")
-    Simple(std::string::String),
-    List(Box<TypeExpr>),
-    Map(Box<TypeExpr>),
-    /// Reference to a resource type (e.g., aws.vpc)
-    Ref(ResourceTypePath),
-    /// Provider-defined schema type (e.g., awscc.ec2.VpcId, awscc.ec2.SubnetId)
-    /// Distinguished from Ref by having a PascalCase final segment.
-    SchemaType {
-        /// Provider name (e.g., "awscc")
-        provider: String,
-        /// Service/namespace path (e.g., "ec2")
-        path: String,
-        /// Type name in PascalCase (e.g., "VpcId")
-        type_name: String,
-    },
-    /// Structural record type: `struct { name: type, ... }`.
-    ///
-    /// Field order matches source order and participates in `PartialEq` —
-    /// two struct types with the same fields in different order are not
-    /// equal. A `Value::Map` satisfies a struct type when every field
-    /// name appears as a key with a value that matches the field's type,
-    /// with no extra keys.
-    Struct {
-        fields: Vec<(String, TypeExpr)>,
-    },
-}
-
-impl std::fmt::Display for TypeExpr {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TypeExpr::String => write!(f, "String"),
-            TypeExpr::Bool => write!(f, "Bool"),
-            TypeExpr::Int => write!(f, "Int"),
-            TypeExpr::Float => write!(f, "Float"),
-            TypeExpr::Simple(name) => write!(f, "{}", snake_to_pascal(name)),
-            TypeExpr::List(inner) => write!(f, "list({})", inner),
-            TypeExpr::Map(inner) => write!(f, "map({})", inner),
-            TypeExpr::Ref(path) => write!(f, "{}", path),
-            TypeExpr::SchemaType {
-                provider,
-                path,
-                type_name,
-            } => write!(f, "{}.{}.{}", provider, path, type_name),
-            TypeExpr::Struct { fields } => {
-                if fields.is_empty() {
-                    write!(f, "struct {{}}")
-                } else {
-                    write!(f, "struct {{ ")?;
-                    for (i, (name, ty)) in fields.iter().enumerate() {
-                        if i > 0 {
-                            write!(f, ", ")?;
-                        }
-                        write!(f, "{}: {}", name, ty)?;
-                    }
-                    write!(f, " }}")
-                }
-            }
-        }
-    }
-}
-
-/// Argument parameter definition (in `arguments { ... }` block)
-#[derive(Debug, Clone)]
-pub struct ArgumentParameter {
-    pub name: String,
-    pub type_expr: TypeExpr,
-    pub default: Option<Value>,
-    /// Optional description (from block form)
-    pub description: Option<String>,
-    /// Optional validation blocks (from block form). Multiple blocks are allowed.
-    pub validations: Vec<ValidationBlock>,
-}
-
-/// A validate block: `validation { condition = <expr> error_message = "..." }`
-#[derive(Debug, Clone)]
-pub struct ValidationBlock {
-    pub condition: ValidateExpr,
-    pub error_message: Option<String>,
-}
-
-// `CompareOp` lives in `expressions::validate_expr` (re-exported above).
-
-/// Validate expression AST node
-#[derive(Debug, Clone, PartialEq)]
-pub enum ValidateExpr {
-    /// Boolean literal
-    Bool(bool),
-    /// Integer literal
-    Int(i64),
-    /// Float literal
-    Float(f64),
-    /// String literal
-    String(String),
-    /// Variable reference (argument name)
-    Var(String),
-    /// Comparison: lhs op rhs
-    Compare {
-        lhs: Box<ValidateExpr>,
-        op: CompareOp,
-        rhs: Box<ValidateExpr>,
-    },
-    /// Logical AND
-    And(Box<ValidateExpr>, Box<ValidateExpr>),
-    /// Logical OR
-    Or(Box<ValidateExpr>, Box<ValidateExpr>),
-    /// Logical NOT
-    Not(Box<ValidateExpr>),
-    /// Function call (e.g., len(x))
-    FunctionCall {
-        name: String,
-        args: Vec<ValidateExpr>,
-    },
-    /// Null literal
-    Null,
-}
-
-/// A require block: `require <condition>, "error message"`
-/// Used for cross-argument constraints at the module top level.
-#[derive(Debug, Clone)]
-pub struct RequireBlock {
-    pub condition: ValidateExpr,
-    pub error_message: String,
-}
-
-/// Attribute parameter definition (in `attributes { ... }` block)
-#[derive(Debug, Clone)]
-pub struct AttributeParameter {
-    pub name: String,
-    pub type_expr: Option<TypeExpr>,
-    pub value: Option<Value>,
-}
-
-/// An export parameter in an `exports { }` block.
-/// Publishes a named value for `upstream_state` consumers.
-#[derive(Debug, Clone, PartialEq)]
-pub struct ExportParameter {
-    pub name: String,
-    pub type_expr: Option<TypeExpr>,
-    pub value: Option<Value>,
-}
-
-/// State manipulation block (import, removed, moved)
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum StateBlock {
-    /// Import existing infrastructure into Carina management
-    Import {
-        /// Target resource address
-        to: ResourceId,
-        /// Cloud provider identifier (e.g., "vpc-0abc123def456")
-        id: String,
-    },
-    /// Remove a resource from state without destroying it
-    Removed {
-        /// Resource address to remove from state
-        from: ResourceId,
-    },
-    /// Rename/move a resource in state without destroy/recreate
-    Moved {
-        /// Old resource address
-        from: ResourceId,
-        /// New resource address
-        to: ResourceId,
-    },
-}
-
-/// Module `use` statement (previously `import`).
-#[derive(Debug, Clone)]
-pub struct UseStatement {
-    pub path: String,
-    pub alias: String,
-}
-
-/// Parameter for a user-defined function
-#[derive(Debug, Clone)]
-pub struct FnParam {
-    pub name: String,
-    pub param_type: Option<TypeExpr>,
-    pub default: Option<Value>,
-}
-
-/// The body of a user-defined function: a value expression.
-/// Functions are pure value transformations only.
-#[derive(Debug, Clone)]
-pub struct UserFunctionBody(pub Value);
-
-/// User-defined pure function
-#[derive(Debug, Clone)]
-pub struct UserFunction {
-    pub name: String,
-    pub params: Vec<FnParam>,
-    /// Optional return type annotation
-    pub return_type: Option<TypeExpr>,
-    /// Local let bindings inside the function body (name, expression)
-    pub local_lets: Vec<(String, Value)>,
-    /// The body of the function
-    pub body: UserFunctionBody,
-}
-
-/// Module call (instantiation)
-#[derive(Debug, Clone)]
-pub struct ModuleCall {
-    pub module_name: String,
-    pub binding_name: Option<String>,
-    pub arguments: HashMap<String, Value>,
-}
-
-/// Provider configuration
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub struct ProviderConfig {
-    pub name: String,
-    pub attributes: IndexMap<String, Value>,
-    /// Default tags to apply to all resources that support tags.
-    /// Extracted from `default_tags = { ... }` in the provider block.
-    #[serde(default, skip_serializing_if = "IndexMap::is_empty")]
-    pub default_tags: IndexMap<String, Value>,
-    /// Provider source (e.g., "github.com/carina-rs/carina-provider-awscc" or "file:///path/to/binary").
-    /// Extracted from the provider block and not passed to the provider.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub source: Option<String>,
-    /// Provider version constraint (e.g., "~0.5.0", "^1.2.0").
-    /// Extracted from the provider block and not passed to the provider.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub version: Option<VersionConstraint>,
-    /// Git revision (branch, tag, or commit SHA) to resolve the provider from CI artifacts.
-    /// Mutually exclusive with `version`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub revision: Option<String>,
-}
-
-/// Backend configuration for state storage
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct BackendConfig {
-    /// Backend type (e.g., "s3", "gcs", "local")
-    pub backend_type: String,
-    /// Backend-specific attributes
-    pub attributes: HashMap<String, Value>,
-}
-
-/// Upstream state reference: `let <binding> = upstream_state { source = "<dir>" }`.
-///
-/// Declares a read-only reference to another Carina configuration's state.
-///
-/// `source` is a directory path. Carina itself is directory-scoped, so every
-/// sibling `.crn` file in a project (or module) shares the same base
-/// directory; the resolution rule is simply:
-///
-/// - Absolute paths are used as-is.
-/// - Relative paths are resolved against the **enclosing project or module
-///   directory** — the one passed to `carina validate` / `plan` / `apply`,
-///   or the module directory containing the `upstream_state` declaration.
-///
-/// Which specific `.crn` file inside that directory declares the
-/// `upstream_state` does not affect resolution. The upstream's backend and
-/// state file are derived from the upstream configuration itself.
-#[derive(Debug, Clone)]
-pub struct UpstreamState {
-    /// The binding name (e.g., "orgs")
-    pub binding: String,
-    /// Source directory (raw, unresolved path)
-    pub source: std::path::PathBuf,
-}
-
-/// Parse result
-#[derive(Debug, Clone, Default)]
-pub struct ParsedFile {
-    pub providers: Vec<ProviderConfig>,
-    pub resources: Vec<Resource>,
-    pub variables: IndexMap<String, Value>,
-    /// Module `use` statements
-    pub uses: Vec<UseStatement>,
-    /// Module calls (instantiations)
-    pub module_calls: Vec<ModuleCall>,
-    /// Top-level argument parameters (directory-based module style)
-    pub arguments: Vec<ArgumentParameter>,
-    /// Top-level attribute parameters (directory-based module style)
-    pub attribute_params: Vec<AttributeParameter>,
-    /// Top-level export parameters (published to upstream_state consumers)
-    pub export_params: Vec<ExportParameter>,
-    /// Backend configuration for state storage
-    pub backend: Option<BackendConfig>,
-    /// State manipulation blocks (import, removed, moved)
-    pub state_blocks: Vec<StateBlock>,
-    /// User-defined pure functions
-    pub user_functions: HashMap<String, UserFunction>,
-    /// Upstream state references (read-only views of other Carina configurations)
-    pub upstream_states: Vec<UpstreamState>,
-    /// Require blocks (cross-argument constraints)
-    pub requires: Vec<RequireBlock>,
-    /// Binding names that are structurally required (if/for/read expressions)
-    /// and should not trigger unused-binding warnings.
-    pub structural_bindings: HashSet<String>,
-    /// Non-fatal warnings collected during parsing.
-    pub warnings: Vec<ParseWarning>,
-    /// For-expressions whose iterables are unresolved; displayed as deferred in plan.
-    pub deferred_for_expressions: Vec<DeferredForExpression>,
-}
-
-impl ParsedFile {
-    /// Iterate every resource reachable from the parsed file — both
-    /// top-level `resources` and the `template_resource` of each deferred
-    /// for-expression — tagged with its origin context.
-    ///
-    /// Per-attribute checkers (type, enum, required, ref validity, etc.)
-    /// should prefer this over `self.resources.iter()` so they stay in sync
-    /// with for-body code. See
-    /// `docs/specs/2026-04-19-unify-resource-walk-design.md` for the
-    /// rationale.
-    pub fn iter_all_resources(&self) -> impl Iterator<Item = (ResourceContext<'_>, &Resource)> {
-        self.resources
-            .iter()
-            .map(|r| (ResourceContext::Direct, r))
-            .chain(
-                self.deferred_for_expressions
-                    .iter()
-                    .map(|d| (ResourceContext::Deferred(d), &d.template_resource)),
-            )
-    }
-
-    /// Find a resource by resource type and name attribute value
-    pub fn find_resource_by_attr(
-        &self,
-        resource_type: &str,
-        attr_name: &str,
-        attr_value: &str,
-    ) -> Option<&Resource> {
-        self.resources.iter().find(|r| {
-            r.id.resource_type == resource_type
-                && matches!(r.get_attr(attr_name), Some(Value::String(n)) if n == attr_value)
-        })
-    }
-
-    /// Print all collected warnings to stderr.
-    pub fn print_warnings(&self) {
-        for w in &self.warnings {
-            let location = match &w.file {
-                Some(f) => format!("{}:{}", f, w.line),
-                None => format!("line {}", w.line),
-            };
-            eprintln!("  ⚠ {}: {}", location, w.message);
-        }
-    }
-
-    /// Expand deferred for-expressions using loaded upstream_state bindings.
-    ///
-    /// For each deferred for-expression whose iterable can now be resolved
-    /// from `remote_bindings`, expand the template into concrete resources
-    /// and add them to `self.resources`. Resolved entries are removed from
-    /// `deferred_for_expressions`; unresolved ones remain (with their
-    /// warning preserved).
-    pub fn expand_deferred_for_expressions(
-        &mut self,
-        remote_bindings: &HashMap<String, HashMap<String, Value>>,
-    ) {
-        let mut expanded_resources = Vec::new();
-        let mut resolved_indices = Vec::new();
-        // Indices where the iterable resolved but had the wrong shape. These
-        // entries stay deferred (user must fix), but their parse-time "not yet
-        // available" warning is replaced by the more specific shape-mismatch
-        // warning collected in `new_warnings`.
-        let mut mismatched_indices: Vec<usize> = Vec::new();
-        let mut new_warnings: Vec<ParseWarning> = Vec::new();
-
-        for (idx, deferred) in self.deferred_for_expressions.iter().enumerate() {
-            // Look up the iterable value in remote_bindings
-            let iterable = remote_bindings
-                .get(&deferred.iterable_binding)
-                .and_then(|attrs| attrs.get(&deferred.iterable_attr));
-
-            let Some(iterable_value) = iterable else {
-                continue;
-            };
-
-            match (&deferred.binding, iterable_value) {
-                // Simple binding: only the value var is bound
-                (ForBinding::Simple(_), Value::List(items)) => {
-                    for (i, item) in items.iter().enumerate() {
-                        let address = format!("{}[{}]", deferred.binding_name, i);
-                        let mut resource = deferred.template_resource.clone();
-                        resource.id.set_name(address.clone());
-                        resource.binding = Some(address);
-                        for (_k, expr) in resource.attributes.iter_mut() {
-                            substitute_placeholder(&mut expr.0, None, None, item);
-                        }
-                        expanded_resources.push(resource);
-                    }
-                    resolved_indices.push(idx);
-                }
-                // Indexed binding: both index and value vars are bound
-                (ForBinding::Indexed(_, _), Value::List(items)) => {
-                    for (i, item) in items.iter().enumerate() {
-                        let address = format!("{}[{}]", deferred.binding_name, i);
-                        let mut resource = deferred.template_resource.clone();
-                        resource.id.set_name(address.clone());
-                        resource.binding = Some(address);
-                        for (_k, expr) in resource.attributes.iter_mut() {
-                            substitute_placeholder(&mut expr.0, Some(i as i64), None, item);
-                        }
-                        expanded_resources.push(resource);
-                    }
-                    resolved_indices.push(idx);
-                }
-                // Map binding expands over maps, substituting both key and value vars
-                (ForBinding::Map(_, _), Value::Map(map)) => {
-                    let mut keys: Vec<&String> = map.keys().collect();
-                    keys.sort();
-                    for key in keys {
-                        let val = &map[key];
-                        let address = format!("{}[\"{}\"]", deferred.binding_name, key);
-                        let mut resource = deferred.template_resource.clone();
-                        resource.id.set_name(address.clone());
-                        resource.binding = Some(address);
-                        for (_k, expr) in resource.attributes.iter_mut() {
-                            substitute_placeholder(&mut expr.0, None, Some(key), val);
-                        }
-                        expanded_resources.push(resource);
-                    }
-                    resolved_indices.push(idx);
-                }
-                // Shape mismatch: replace the original "not yet available" warning
-                // with a specific shape-mismatch warning, leave entry deferred.
-                (ForBinding::Map(_, _), Value::List(_)) => {
-                    mismatched_indices.push(idx);
-                    new_warnings.push(ParseWarning {
-                        file: deferred.file.clone(),
-                        line: deferred.line,
-                        message: format!(
-                            "for binding expected map iterable but `{}.{}` resolved to a list. \
-                             Fix either the upstream export shape or the downstream binding.",
-                            deferred.iterable_binding, deferred.iterable_attr,
-                        ),
-                    });
-                }
-                (ForBinding::Simple(_), Value::Map(_))
-                | (ForBinding::Indexed(_, _), Value::Map(_)) => {
-                    mismatched_indices.push(idx);
-                    new_warnings.push(ParseWarning {
-                        file: deferred.file.clone(),
-                        line: deferred.line,
-                        message: format!(
-                            "for binding expected list iterable but `{}.{}` resolved to a map. \
-                             Fix either the upstream export shape or the downstream binding.",
-                            deferred.iterable_binding, deferred.iterable_attr,
-                        ),
-                    });
-                }
-                _ => {
-                    // Iterable is not a list or map — leave deferred
-                }
-            }
-        }
-
-        // Remove parse-time warnings for mismatched entries (the new
-        // shape-mismatch warning replaces the generic "not yet available" one).
-        for idx in &mismatched_indices {
-            let deferred = &self.deferred_for_expressions[*idx];
-            self.warnings
-                .retain(|w| w.line != deferred.line || w.file != deferred.file);
-        }
-
-        // Remove resolved deferred entries (reverse order to preserve indices)
-        for idx in resolved_indices.into_iter().rev() {
-            // Also remove the corresponding warning
-            let deferred = &self.deferred_for_expressions[idx];
-            self.warnings
-                .retain(|w| w.line != deferred.line || w.file != deferred.file);
-            self.deferred_for_expressions.remove(idx);
-        }
-
-        self.resources.extend(expanded_resources);
-        self.warnings.extend(new_warnings);
-    }
-}
-
-/// Substitute deferred-for-expression placeholders in a Value tree.
-///
-/// Replaces `DEFERRED_UPSTREAM_PLACEHOLDER` with `value`. If `index` is
-/// supplied (indexed-binding expansion), replaces
-/// `DEFERRED_UPSTREAM_INDEX_PLACEHOLDER` with the integer index. If `key`
-/// is supplied (map-binding expansion), replaces
-/// `DEFERRED_UPSTREAM_KEY_PLACEHOLDER` with the key string. Recurses into
-/// all compound Value variants so placeholders nested inside
-/// interpolations / function calls / secrets are reached.
-fn substitute_placeholder(v: &mut Value, index: Option<i64>, key: Option<&str>, value: &Value) {
-    match v {
-        Value::String(s) if s == DEFERRED_UPSTREAM_PLACEHOLDER => {
-            *v = value.clone();
-        }
-        Value::String(s) if s == DEFERRED_UPSTREAM_KEY_PLACEHOLDER => {
-            if let Some(k) = key {
-                *v = Value::String(k.to_string());
-            }
-        }
-        Value::String(s) if s == DEFERRED_UPSTREAM_INDEX_PLACEHOLDER => {
-            if let Some(i) = index {
-                *v = Value::Int(i);
-            }
-        }
-        Value::List(items) => {
-            for item in items.iter_mut() {
-                substitute_placeholder(item, index, key, value);
-            }
-        }
-        Value::Map(map) => {
-            for val in map.values_mut() {
-                substitute_placeholder(val, index, key, value);
-            }
-        }
-        Value::FunctionCall { args, .. } => {
-            for arg in args.iter_mut() {
-                substitute_placeholder(arg, index, key, value);
-            }
-        }
-        Value::Interpolation(parts) => {
-            for part in parts.iter_mut() {
-                if let crate::resource::InterpolationPart::Expr(inner) = part {
-                    substitute_placeholder(inner, index, key, value);
-                }
-            }
-        }
-        Value::Secret(inner) => {
-            substitute_placeholder(inner, index, key, value);
-        }
-        _ => {}
-    }
-}
-
-/// Parse context (variable scope)
-#[derive(Clone)]
-struct ParseContext<'cfg> {
-    /// Variables bound by `let` statements during parsing. Carries
-    /// `EvalValue` rather than `Value` because partial applications
-    /// (e.g. `let f = join("-")`) produce a closure that lives until a
-    /// later pipe finishes the application. Lowered to `Value` at the
-    /// end of `parse(...)`; an unfinished closure surfaces as a
-    /// parse-time error.
-    variables: IndexMap<String, EvalValue>,
-    /// Resource bindings (binding_name -> Resource)
-    resource_bindings: HashMap<String, Resource>,
-    /// Imported modules (alias -> path)
-    imported_modules: HashMap<String, String>,
-    /// User-defined functions
-    user_functions: HashMap<String, UserFunction>,
-    /// Functions currently being evaluated (for recursion detection)
-    evaluating_functions: Vec<String>,
-    /// Parser configuration (decryptor, custom validators)
-    config: &'cfg ProviderContext,
-    /// Binding names from structurally-required expressions (if/for/read)
-    structural_bindings: HashSet<String>,
-    /// Upstream state bindings (binding_name -> UpstreamState)
-    upstream_states: HashMap<String, UpstreamState>,
-    /// Non-fatal warnings collected during parsing
-    warnings: Vec<ParseWarning>,
-    /// Deferred for-expressions collected during parsing
-    deferred_for_expressions: Vec<DeferredForExpression>,
-}
-
-impl<'cfg> ParseContext<'cfg> {
-    fn new(config: &'cfg ProviderContext) -> Self {
-        Self {
-            variables: IndexMap::new(),
-            resource_bindings: HashMap::new(),
-            imported_modules: HashMap::new(),
-            user_functions: HashMap::new(),
-            evaluating_functions: Vec::new(),
-            config,
-            structural_bindings: HashSet::new(),
-            upstream_states: HashMap::new(),
-            warnings: Vec::new(),
-            deferred_for_expressions: Vec::new(),
-        }
-    }
-
-    fn set_variable(&mut self, name: String, value: impl Into<EvalValue>) {
-        self.variables.insert(name, value.into());
-    }
-
-    fn get_variable(&self, name: &str) -> Option<&EvalValue> {
-        self.variables.get(name)
-    }
-
-    fn set_resource_binding(&mut self, name: String, resource: Resource) {
-        self.resource_bindings.insert(name, resource);
-    }
-
-    fn is_resource_binding(&self, name: &str) -> bool {
-        self.resource_bindings.contains_key(name)
-    }
-}
-
-/// Helper to get the next element from a pest iterator, returning a ParseError on failure
-fn next_pair<'a>(
-    iter: &mut pest::iterators::Pairs<'a, Rule>,
-    expected: &str,
-    context: &str,
-) -> Result<pest::iterators::Pair<'a, Rule>, ParseError> {
-    iter.next().ok_or_else(|| ParseError::InternalError {
-        expected: expected.to_string(),
-        context: context.to_string(),
-    })
-}
-
-/// Extract a key string from either an identifier or a quoted string pair.
-/// For identifiers, returns the raw text. For strings, extracts the content
-/// without quotes (supports both single-quoted and double-quoted strings).
-fn extract_key_string(pair: pest::iterators::Pair<'_, Rule>) -> Result<String, ParseError> {
-    match pair.as_rule() {
-        Rule::identifier => Ok(pair.as_str().to_string()),
-        Rule::string => {
-            let inner = pair
-                .into_inner()
-                .next()
-                .ok_or_else(|| ParseError::InternalError {
-                    expected: "string content".to_string(),
-                    context: "map/attribute key".to_string(),
-                })?;
-            match inner.as_rule() {
-                Rule::single_quoted_string => {
-                    // Extract content between quotes
-                    let content = inner
-                        .into_inner()
-                        .next()
-                        .map(|p| p.as_str().to_string())
-                        .unwrap_or_default();
-                    Ok(content)
-                }
-                Rule::double_quoted_string => {
-                    let mut result = String::new();
-                    for part in inner.into_inner() {
-                        match part.as_rule() {
-                            Rule::string_part => {
-                                let inner_part = part.into_inner().next().unwrap();
-                                match inner_part.as_rule() {
-                                    Rule::string_literal => result.push_str(inner_part.as_str()),
-                                    Rule::interpolation => {
-                                        return Err(ParseError::InternalError {
-                                            expected: "literal string".to_string(),
-                                            context: "interpolation not supported in map keys"
-                                                .to_string(),
-                                        });
-                                    }
-                                    _ => result.push_str(inner_part.as_str()),
-                                }
-                            }
-                            _ => result.push_str(part.as_str()),
-                        }
-                    }
-                    Ok(result)
-                }
-                _ => Ok(inner.as_str().to_string()),
-            }
-        }
-        _ => Ok(pair.as_str().to_string()),
-    }
-}
-
-/// Helper to get the first inner pair from a pest pair
-fn first_inner<'a>(
-    pair: pest::iterators::Pair<'a, Rule>,
-    expected: &str,
-    context: &str,
-) -> Result<pest::iterators::Pair<'a, Rule>, ParseError> {
-    pair.into_inner()
-        .next()
-        .ok_or_else(|| ParseError::InternalError {
-            expected: expected.to_string(),
-            context: context.to_string(),
-        })
-}
-
-/// Parse a .crn file with the given configuration.
-///
-/// The config allows injecting a decryptor function for `decrypt()` calls
-/// and custom type validators from provider crates.
-pub fn parse(input: &str, config: &ProviderContext) -> Result<ParsedFile, ParseError> {
-    let preprocess_result =
-        crate::heredoc::preprocess_heredocs(input).map_err(|e| ParseError::InvalidExpression {
-            line: 0,
-            message: e.to_string(),
-        })?;
-    let pairs = CarinaParser::parse(Rule::file, &preprocess_result.source)?;
-
-    let mut ctx = ParseContext::new(config);
-    let mut providers = Vec::new();
-    let mut resources = Vec::new();
-    let mut uses = Vec::new();
-    let mut module_calls = Vec::new();
-    let mut arguments = Vec::new();
-    let mut attribute_params = Vec::new();
-    let mut export_params = Vec::new();
-    let mut backend = None;
-    let mut state_blocks = Vec::new();
-    let mut upstream_states: Vec<UpstreamState> = Vec::new();
-    let mut requires = Vec::new();
-    let mut anon_for_counter = 0usize;
-    let mut anon_if_counter = 0usize;
-
-    for pair in pairs {
-        if pair.as_rule() == Rule::file {
-            for inner in pair.into_inner() {
-                if inner.as_rule() == Rule::statement {
-                    for stmt in inner.into_inner() {
-                        match stmt.as_rule() {
-                            Rule::backend_block => {
-                                backend = Some(parse_backend_block(stmt, &ctx)?);
-                            }
-                            Rule::provider_block => {
-                                let provider = parse_provider_block(stmt, &ctx)?;
-                                providers.push(provider);
-                            }
-                            Rule::arguments_block => {
-                                let parsed_arguments =
-                                    parse_arguments_block(stmt, config, &mut ctx.warnings)?;
-                                for arg in &parsed_arguments {
-                                    // Register argument names as lexical bindings so that
-                                    // `vpc.vpc_id` resolves as ResourceRef and `cidr_block`
-                                    // resolves as a variable reference during parsing.
-                                    // No `arguments.` prefix needed.
-                                    let placeholder_ref = Value::resource_ref(
-                                        arg.name.clone(),
-                                        String::new(),
-                                        vec![],
-                                    );
-                                    ctx.set_variable(arg.name.clone(), placeholder_ref);
-                                    let placeholder = Resource::new("_argument", &arg.name);
-                                    ctx.set_resource_binding(arg.name.clone(), placeholder);
-                                }
-                                arguments.extend(parsed_arguments);
-                            }
-                            Rule::attributes_block => {
-                                let parsed_attribute_params = {
-                                    let warnings = std::mem::take(&mut ctx.warnings);
-                                    let mut warnings = warnings;
-                                    let result = parse_attributes_block(stmt, &ctx, &mut warnings);
-                                    ctx.warnings = warnings;
-                                    result?
-                                };
-                                attribute_params.extend(parsed_attribute_params);
-                            }
-                            Rule::exports_block => {
-                                let parsed_export_params = {
-                                    let warnings = std::mem::take(&mut ctx.warnings);
-                                    let mut warnings = warnings;
-                                    let result = parse_exports_block(stmt, &ctx, &mut warnings);
-                                    ctx.warnings = warnings;
-                                    result?
-                                };
-                                export_params.extend(parsed_export_params);
-                            }
-                            Rule::import_state_block => {
-                                state_blocks.push(parse_import_state_block(stmt)?);
-                            }
-                            Rule::removed_block => {
-                                state_blocks.push(parse_removed_block(stmt)?);
-                            }
-                            Rule::moved_block => {
-                                state_blocks.push(parse_moved_block(stmt)?);
-                            }
-                            Rule::require_statement => {
-                                requires.push(parse_require_statement(stmt)?);
-                            }
-                            Rule::for_expr => {
-                                let iterable_name =
-                                    extract_for_iterable_name(&stmt, anon_for_counter);
-                                anon_for_counter += 1;
-                                let (expanded_resources, expanded_module_calls) =
-                                    parse_for_expr(stmt, &mut ctx, &iterable_name)?;
-                                resources.extend(expanded_resources);
-                                module_calls.extend(expanded_module_calls);
-                            }
-                            Rule::if_expr => {
-                                let binding_name = format!("_if{}", anon_if_counter);
-                                anon_if_counter += 1;
-                                let (_value, expanded_resources, expanded_module_calls, _import) =
-                                    parse_if_expr(stmt, &mut ctx, &binding_name)?;
-                                resources.extend(expanded_resources);
-                                module_calls.extend(expanded_module_calls);
-                            }
-                            Rule::fn_def => {
-                                let user_fn = {
-                                    let warnings = std::mem::take(&mut ctx.warnings);
-                                    let mut warnings = warnings;
-                                    let result = parse_fn_def(stmt, &ctx, &mut warnings);
-                                    ctx.warnings = warnings;
-                                    result?
-                                };
-                                let fn_name = user_fn.name.clone();
-                                // Check for shadowing builtins
-                                if crate::builtins::evaluate_builtin(&fn_name, &[]).is_ok()
-                                    || crate::builtins::builtin_functions()
-                                        .iter()
-                                        .any(|f| f.name == fn_name)
-                                {
-                                    return Err(ParseError::UserFunctionError(format!(
-                                        "function '{fn_name}' shadows a built-in function"
-                                    )));
-                                }
-                                if ctx.user_functions.contains_key(&fn_name) {
-                                    return Err(ParseError::UserFunctionError(format!(
-                                        "duplicate function definition: '{fn_name}'"
-                                    )));
-                                }
-                                ctx.user_functions.insert(fn_name, user_fn);
-                            }
-                            Rule::let_binding => {
-                                let (line, _) = stmt.as_span().start_pos().line_col();
-                                let (
-                                    name,
-                                    value,
-                                    expanded_resources,
-                                    expanded_module_calls,
-                                    maybe_import,
-                                    is_structural,
-                                ) = parse_let_binding_extended(stmt, &mut ctx)?;
-                                if is_structural {
-                                    ctx.structural_bindings.insert(name.clone());
-                                }
-                                let is_discard = name == "_";
-                                let is_upstream_state = ctx.upstream_states.contains_key(&name);
-                                if !is_discard {
-                                    if ctx.variables.contains_key(&name)
-                                        || ctx.resource_bindings.contains_key(&name)
-                                    {
-                                        return Err(ParseError::DuplicateBinding { name, line });
-                                    }
-                                    if !is_upstream_state {
-                                        ctx.set_variable(name.clone(), value);
-                                    }
-                                }
-                                if !expanded_resources.is_empty() {
-                                    if !is_discard {
-                                        // Register the binding name as a resource binding
-                                        // (use the first resource as placeholder)
-                                        ctx.set_resource_binding(
-                                            name.clone(),
-                                            expanded_resources[0].clone(),
-                                        );
-                                    }
-                                    resources.extend(expanded_resources);
-                                }
-                                if !expanded_module_calls.is_empty() {
-                                    for mut call in expanded_module_calls {
-                                        if call.binding_name.is_none() {
-                                            call.binding_name = Some(name.clone());
-                                        }
-                                        module_calls.push(call);
-                                    }
-                                    if !is_discard {
-                                        // Register as a resource binding so that
-                                        // `name.attr` resolves as ResourceRef
-                                        let placeholder = Resource::new("_module_binding", &name);
-                                        ctx.set_resource_binding(name.clone(), placeholder);
-                                    }
-                                }
-                                if is_upstream_state && !is_discard {
-                                    let placeholder = Resource::new("_upstream_state", &name);
-                                    ctx.set_resource_binding(name.clone(), placeholder);
-                                    upstream_states.push(ctx.upstream_states[&name].clone());
-                                }
-                                if let Some(use_stmt) = maybe_import {
-                                    ctx.imported_modules
-                                        .insert(use_stmt.alias.clone(), use_stmt.path.clone());
-                                    uses.push(use_stmt);
-                                }
-                            }
-                            Rule::module_call => {
-                                let call = parse_module_call(stmt, &ctx)?;
-                                module_calls.push(call);
-                            }
-                            Rule::anonymous_resource => {
-                                let resource = parse_anonymous_resource(stmt, &ctx)?;
-                                resources.push(resource);
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // Second pass: resolve forward references.
-    // During parsing, unknown 2-part identifiers (e.g., vpc.vpc_id where vpc is
-    // declared later) become String values like "vpc.vpc_id". Now that we have the
-    // full binding set, convert matching ones to ResourceRef.
-    resolve_forward_references(
-        &ctx.resource_bindings,
-        &mut resources,
-        &mut attribute_params,
-        &mut module_calls,
-        &mut export_params,
-    );
-
-    // "Is every ResourceRef root declared somewhere?" is a semantic
-    // question the per-file parse cannot answer: the referent may live
-    // in a sibling `.crn`. The check runs post-merge via
-    // `check_identifier_scope(&ParsedFile)` — wired into
-    // `load_configuration_with_config`. See #2126 / #2138.
-
-    // Lower the evaluator-internal `EvalValue` bindings to user-facing
-    // `Value`. Closure bindings are dropped: they are evaluator-only
-    // artifacts (partial applications produced by `let f = builtin(x)`
-    // and consumed by a later pipe like `data |> f()`). After the parse
-    // pass finishes, only fully-reduced `Value`s belong in
-    // `ParsedFile.variables`; nothing downstream knows how to handle a
-    // closure. Pipe / call paths read from `ctx.variables` directly via
-    // `get_variable`, so the closure was already available where it
-    // mattered.
-    let variables: IndexMap<String, Value> = ctx
-        .variables
-        .into_iter()
-        .filter_map(|(name, eval)| match eval.into_value() {
-            Ok(v) => Some((name, v)),
-            Err(_leak) => None,
-        })
-        .collect();
-
-    Ok(ParsedFile {
-        providers,
-        resources,
-        variables,
-        uses,
-        module_calls,
-        arguments,
-        attribute_params,
-        export_params,
-        backend,
-        state_blocks,
-        user_functions: ctx.user_functions,
-        upstream_states,
-        requires,
-        structural_bindings: ctx.structural_bindings,
-        warnings: ctx.warnings,
-        deferred_for_expressions: ctx.deferred_for_expressions,
-    })
-}
-
-/// Parse arguments block
-fn parse_arguments_block(
-    pair: pest::iterators::Pair<Rule>,
-    config: &ProviderContext,
-    warnings: &mut Vec<ParseWarning>,
-) -> Result<Vec<ArgumentParameter>, ParseError> {
-    let mut arguments = Vec::new();
-    let ctx = ParseContext::new(config);
-
-    for param in pair.into_inner() {
-        if param.as_rule() == Rule::arguments_param {
-            let mut param_inner = param.into_inner();
-            let name = next_pair(&mut param_inner, "parameter name", "arguments block")?
-                .as_str()
-                .to_string();
-            let type_expr = parse_type_expr(
-                next_pair(&mut param_inner, "type expression", "arguments parameter")?,
-                config,
-                warnings,
-            )?;
-
-            // Check if the next element is a block form or simple default
-            if let Some(next) = param_inner.next() {
-                if next.as_rule() == Rule::arguments_param_block {
-                    // Block form: parse description, default, validate blocks from attrs
-                    let mut description = None;
-                    let mut default = None;
-                    let mut validations = Vec::new();
-                    for attr in next.into_inner() {
-                        if attr.as_rule() == Rule::arguments_param_attr {
-                            let inner_attr =
-                                first_inner(attr, "attribute", "arguments_param_attr")?;
-                            match inner_attr.as_rule() {
-                                Rule::arg_description_attr => {
-                                    let string_pair =
-                                        first_inner(inner_attr, "string", "arg_description_attr")?;
-                                    let value = parse_string_value(string_pair, &ctx)?;
-                                    if let Value::String(s) = value {
-                                        description = Some(s);
-                                    }
-                                }
-                                Rule::arg_default_attr => {
-                                    let expr_pair =
-                                        first_inner(inner_attr, "expression", "arg_default_attr")?;
-                                    default = Some(parse_expression(expr_pair, &ctx)?);
-                                }
-                                Rule::arg_validation_block => {
-                                    let mut rule = None;
-                                    let mut error_msg = None;
-                                    for vattr in inner_attr.into_inner() {
-                                        if vattr.as_rule() == Rule::validation_block_attr {
-                                            let inner_vattr = first_inner(
-                                                vattr,
-                                                "validation_block_attr",
-                                                "validation_block_attr",
-                                            )?;
-                                            match inner_vattr.as_rule() {
-                                                Rule::validation_condition_attr => {
-                                                    let validate_pair = first_inner(
-                                                        inner_vattr,
-                                                        "validate_expr",
-                                                        "validation_condition_attr",
-                                                    )?;
-                                                    rule =
-                                                        Some(parse_validate_expr(validate_pair)?);
-                                                }
-                                                Rule::validation_error_message_attr => {
-                                                    let string_pair = first_inner(
-                                                        inner_vattr,
-                                                        "string",
-                                                        "validation_error_message_attr",
-                                                    )?;
-                                                    let value =
-                                                        parse_string_value(string_pair, &ctx)?;
-                                                    if let Value::String(s) = value {
-                                                        error_msg = Some(s);
-                                                    }
-                                                }
-                                                _ => {}
-                                            }
-                                        }
-                                    }
-                                    if let Some(condition) = rule {
-                                        validations.push(ValidationBlock {
-                                            condition,
-                                            error_message: error_msg,
-                                        });
-                                    }
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-                    arguments.push(ArgumentParameter {
-                        name,
-                        type_expr,
-                        default,
-                        description,
-                        validations,
-                    });
-                } else {
-                    // Simple form: the next element is the default expression
-                    let default = Some(parse_expression(next, &ctx)?);
-                    arguments.push(ArgumentParameter {
-                        name,
-                        type_expr,
-                        default,
-                        description: None,
-                        validations: Vec::new(),
-                    });
-                }
-            } else {
-                // No default, no block
-                arguments.push(ArgumentParameter {
-                    name,
-                    type_expr,
-                    default: None,
-                    description: None,
-                    validations: Vec::new(),
-                });
-            }
-        }
-    }
-
-    Ok(arguments)
-}
-
-/// Parse a require statement: `require <validate_expr>, "error message"`
-fn parse_require_statement(pair: pest::iterators::Pair<Rule>) -> Result<RequireBlock, ParseError> {
-    let mut inner = pair.into_inner();
-    let condition_pair = next_pair(&mut inner, "validate_expr", "require statement")?;
-    let condition = parse_validate_expr(condition_pair)?;
-    let message_pair = next_pair(&mut inner, "string", "require statement")?;
-    let raw = message_pair.as_str();
-    let error_message = raw[1..raw.len() - 1].to_string();
-    Ok(RequireBlock {
-        condition,
-        error_message,
-    })
-}
-
-/// Parse attributes block
-fn parse_attributes_block(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    warnings: &mut Vec<ParseWarning>,
-) -> Result<Vec<AttributeParameter>, ParseError> {
-    let mut attribute_params = Vec::new();
-
-    for param in pair.into_inner() {
-        if param.as_rule() == Rule::attributes_param {
-            let mut param_inner = param.into_inner();
-            let name = next_pair(&mut param_inner, "parameter name", "attributes block")?
-                .as_str()
-                .to_string();
-
-            // Check whether the next inner pair is a type_expr or an expression
-            let next = next_pair(
-                &mut param_inner,
-                "type or expression",
-                "attributes parameter",
-            )?;
-            let (type_expr, value) = if next.as_rule() == Rule::type_expr {
-                // Has explicit type annotation: name: type = expr
-                let type_expr = Some(parse_type_expr(next, ctx.config, warnings)?);
-                let expr = next_pair(&mut param_inner, "value expression", "attributes parameter")?;
-                let value = Some(parse_expression(expr, ctx)?);
-                (type_expr, value)
-            } else {
-                // No type annotation: name = expr
-                let value = Some(parse_expression(next, ctx)?);
-                (None, value)
-            };
-
-            attribute_params.push(AttributeParameter {
-                name,
-                type_expr,
-                value,
-            });
-        }
-    }
-
-    Ok(attribute_params)
-}
-
-fn parse_exports_block(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    warnings: &mut Vec<ParseWarning>,
-) -> Result<Vec<ExportParameter>, ParseError> {
-    let mut export_params = Vec::new();
-
-    for param in pair.into_inner() {
-        if param.as_rule() == Rule::exports_param {
-            let mut param_inner = param.into_inner();
-            let name = next_pair(&mut param_inner, "parameter name", "exports block")?
-                .as_str()
-                .to_string();
-
-            let next = next_pair(&mut param_inner, "type or expression", "exports parameter")?;
-            let (type_expr, value) = if next.as_rule() == Rule::type_expr {
-                let type_expr = Some(parse_type_expr(next, ctx.config, warnings)?);
-                let expr = next_pair(&mut param_inner, "value expression", "exports parameter")?;
-                let value = Some(parse_expression(expr, ctx)?);
-                (type_expr, value)
-            } else {
-                let value = Some(parse_expression(next, ctx)?);
-                (None, value)
-            };
-
-            export_params.push(ExportParameter {
-                name,
-                type_expr,
-                value,
-            });
-        }
-    }
-
-    Ok(export_params)
-}
-
-/// Parse type expression
-fn parse_type_expr(
-    pair: pest::iterators::Pair<Rule>,
-    config: &ProviderContext,
-    warnings: &mut Vec<ParseWarning>,
-) -> Result<TypeExpr, ParseError> {
-    let _ = warnings;
-    let inner = first_inner(pair, "type", "type expression")?;
-    match inner.as_rule() {
-        Rule::type_simple => {
-            let line = inner.as_span().start_pos().line_col().0;
-            let text = inner.as_str();
-            match text {
-                "String" => Ok(TypeExpr::String),
-                "Bool" => Ok(TypeExpr::Bool),
-                "Int" => Ok(TypeExpr::Int),
-                "Float" => Ok(TypeExpr::Float),
-                // Phase C: the transition window for snake_case primitives
-                // and custom types has closed. The parser accepts only
-                // PascalCase type names (naming-conventions design D1).
-                "string" | "bool" | "int" | "float" => Err(ParseError::InvalidExpression {
-                    line,
-                    message: format!(
-                        "unknown type '{text}'; primitive types are PascalCase — use '{}' instead",
-                        snake_to_pascal(text)
-                    ),
-                }),
-                other if other.chars().next().is_some_and(|c| c.is_ascii_uppercase()) => {
-                    Ok(TypeExpr::Simple(pascal_to_snake(other)))
-                }
-                other => Err(ParseError::InvalidExpression {
-                    line,
-                    message: format!(
-                        "unknown type '{other}'; custom types are PascalCase — use '{}' instead",
-                        snake_to_pascal(other)
-                    ),
-                }),
-            }
-        }
-        Rule::type_generic => {
-            // Get the full string representation to determine if it's list or map
-            let full_str = inner.as_str();
-            let is_list = full_str.starts_with("list");
-
-            // Get the inner type expression
-            let mut generic_inner = inner.into_inner();
-            let inner_type = parse_type_expr(
-                next_pair(&mut generic_inner, "inner type", "generic type expression")?,
-                config,
-                warnings,
-            )?;
-
-            if is_list {
-                Ok(TypeExpr::List(Box::new(inner_type)))
-            } else {
-                Ok(TypeExpr::Map(Box::new(inner_type)))
-            }
-        }
-        Rule::type_ref => {
-            // Parse resource_type_path directly (e.g., aws.vpc or awscc.ec2.VpcId)
-            let mut ref_inner = inner.into_inner();
-            let path_str = next_pair(&mut ref_inner, "resource type path", "type ref")?.as_str();
-            let parts: Vec<&str> = path_str.split('.').collect();
-
-            // A 3+ segment path with a PascalCase final segment is ambiguous:
-            // `aws.ec2.Vpc` is a resource kind (Ref), `awscc.ec2.VpcId` is a
-            // schema type. Disambiguate by asking the provider context:
-            // registered schema types become SchemaType, everything else
-            // falls back to Ref.
-            let has_pascal_tail = parts.len() >= 3
-                && parts
-                    .last()
-                    .is_some_and(|s| s.starts_with(|c: char| c.is_uppercase()));
-            if has_pascal_tail {
-                let provider = parts[0];
-                let path = parts[1..parts.len() - 1].join(".");
-                let type_name = parts.last().unwrap();
-                if config.is_schema_type(provider, &path, type_name) {
-                    return Ok(TypeExpr::SchemaType {
-                        provider: provider.to_string(),
-                        path,
-                        type_name: type_name.to_string(),
-                    });
-                }
-            }
-            let path = ResourceTypePath::parse(path_str).ok_or_else(|| {
-                ParseError::InvalidResourceType(format!("Invalid resource type path: {}", path_str))
-            })?;
-            Ok(TypeExpr::Ref(path))
-        }
-        Rule::type_struct => {
-            let mut fields: Vec<(String, TypeExpr)> = Vec::new();
-            for child in inner.into_inner() {
-                if child.as_rule() != Rule::struct_field_list {
-                    continue;
-                }
-                for field_pair in child.into_inner() {
-                    if field_pair.as_rule() != Rule::struct_field {
-                        continue;
-                    }
-                    let mut field_inner = field_pair.into_inner();
-                    let name = next_pair(&mut field_inner, "field name", "struct field")?
-                        .as_str()
-                        .to_string();
-                    let ty = parse_type_expr(
-                        next_pair(&mut field_inner, "field type", "struct field")?,
-                        config,
-                        warnings,
-                    )?;
-                    if fields.iter().any(|(existing, _)| existing == &name) {
-                        return Err(ParseError::InvalidResourceType(format!(
-                            "struct has duplicate field name '{name}'"
-                        )));
-                    }
-                    fields.push((name, ty));
-                }
-            }
-            Ok(TypeExpr::Struct { fields })
-        }
-        _ => Ok(TypeExpr::String),
-    }
-}
-
-/// Parse import expression (RHS of `let name = use { source = "path" }`)
-fn parse_use_expr(
-    pair: pest::iterators::Pair<Rule>,
-    binding_name: &str,
-    ctx: &ParseContext,
-) -> Result<UseStatement, ParseError> {
-    let span = pair.as_span();
-    let line = span.start_pos().line_col().0;
-    let mut source: Option<String> = None;
-
-    for attr in pair.into_inner() {
-        if attr.as_rule() != Rule::attribute {
-            continue;
-        }
-        let attr_span = attr.as_span();
-        let attr_line = attr_span.start_pos().line_col().0;
-        let mut attr_inner = attr.into_inner();
-        let key = next_pair(&mut attr_inner, "attribute name", "use expression")?
-            .as_str()
-            .to_string();
-        let value_pair = next_pair(&mut attr_inner, "attribute value", "use expression")?;
-        if key != "source" {
-            return Err(ParseError::InvalidExpression {
-                line: attr_line,
-                message: format!("`use` block only accepts a `source` attribute, got `{key}`"),
-            });
-        }
-        let value = parse_expression(value_pair, ctx)?;
-        let Value::String(path) = value else {
-            return Err(ParseError::InvalidExpression {
-                line: attr_line,
-                message: "`use` block `source` must be a string literal".to_string(),
-            });
-        };
-        if source.is_some() {
-            return Err(ParseError::InvalidExpression {
-                line: attr_line,
-                message: "`use` block has more than one `source` attribute".to_string(),
-            });
-        }
-        source = Some(path);
-    }
-
-    let path = source.ok_or_else(|| ParseError::InvalidExpression {
-        line,
-        message: "`use` block must have a `source` attribute".to_string(),
-    })?;
-
-    Ok(UseStatement {
-        path,
-        alias: binding_name.to_string(),
-    })
-}
-
-/// Parse module call
-fn parse_module_call(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<ModuleCall, ParseError> {
-    let span = pair.as_span();
-    let mut inner = pair.into_inner();
-    let module_name = next_pair(&mut inner, "module name", "module call")?
-        .as_str()
-        .to_string();
-
-    if module_name == "remote_state" {
-        return Err(ParseError::InvalidExpression {
-            line: span.start_pos().line_col().0,
-            message: "`remote_state` has been replaced by `let <binding> = upstream_state { source = \"...\" }`".to_string(),
-        });
-    }
-
-    let mut arguments = HashMap::new();
-    for arg in inner {
-        if arg.as_rule() == Rule::module_call_arg {
-            let mut arg_inner = arg.into_inner();
-            let key = next_pair(&mut arg_inner, "argument name", "module call argument")?
-                .as_str()
-                .to_string();
-            let value = parse_expression(
-                next_pair(&mut arg_inner, "argument value", "module call argument")?,
-                ctx,
-            )?;
-            arguments.insert(key, value);
-        }
-    }
-
-    Ok(ModuleCall {
-        module_name,
-        binding_name: None,
-        arguments,
-    })
-}
-
-/// Result of parsing the RHS of a let binding: (value, resources, module_calls, use_statement)
-/// Tuple returned by the let-binding parser. The RHS is `EvalValue`
-/// rather than `Value` so partial applications (closures) can survive
-/// until a later pipe finishes them; the surrounding parse pass lowers
-/// each binding to `Value` at the end of `parse(...)`.
-type LetBindingRhs = (
-    EvalValue,
-    Vec<Resource>,
-    Vec<ModuleCall>,
-    Option<UseStatement>,
-);
-
-/// Extended parse_let_binding that also handles module calls, imports, and for expressions.
-///
-/// Returns `(name, value, resources, module_calls, import, is_structural)`.
-/// `is_structural` is true when the RHS is an if/for/read expression, meaning the
-/// `let` binding is structurally required and should not trigger unused-binding warnings.
-#[allow(clippy::type_complexity)]
-fn parse_let_binding_extended(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-) -> Result<
-    (
-        String,
-        EvalValue,
-        Vec<Resource>,
-        Vec<ModuleCall>,
-        Option<UseStatement>,
-        bool,
-    ),
-    ParseError,
-> {
-    let mut inner = pair.into_inner();
-    let name = next_pair(&mut inner, "binding name", "let binding")?
-        .as_str()
-        .to_string();
-    let rhs_pair = next_pair(&mut inner, "expression", "let binding")?;
-
-    // `use` is the only meta-statement allowed as a let-binding RHS. The
-    // grammar permits it here and only here (see carina.pest); everywhere
-    // else it is a parse error.
-    if rhs_pair.as_rule() == Rule::use_expr {
-        let use_stmt = parse_use_expr(rhs_pair, &name, ctx)?;
-        let value = Value::String(format!("${{use:{}}}", use_stmt.path));
-        return Ok((
-            name,
-            EvalValue::from_value(value),
-            vec![],
-            vec![],
-            Some(use_stmt),
-            false,
-        ));
-    }
-
-    // Detect if the RHS is a structurally-required expression (if/for/read)
-    let is_structural = detect_structural_rhs(&rhs_pair);
-
-    // Check if it's a module call, resource expression, or for expression
-    let (value, expanded_resources, module_calls, maybe_import) =
-        parse_expression_with_resource_or_module(rhs_pair, ctx, &name)?;
-
-    Ok((
-        name,
-        value,
-        expanded_resources,
-        module_calls,
-        maybe_import,
-        is_structural,
-    ))
-}
-
-/// Detect if an expression pair's innermost primary is an if/for/read/upstream_state expression.
-fn detect_structural_rhs(pair: &pest::iterators::Pair<Rule>) -> bool {
-    // Walk into expression -> pipe_expr -> compose_expr -> primary -> inner
-    fn find_inner_rule(pair: &pest::iterators::Pair<Rule>) -> Option<Rule> {
-        let inner = pair.clone().into_inner().next()?;
-        match inner.as_rule() {
-            Rule::if_expr
-            | Rule::for_expr
-            | Rule::read_resource_expr
-            | Rule::upstream_state_expr => Some(inner.as_rule()),
-            Rule::pipe_expr | Rule::compose_expr | Rule::coalesce_expr | Rule::expression => {
-                find_inner_rule(&inner)
-            }
-            Rule::primary => {
-                let primary_inner = inner.into_inner().next()?;
-                match primary_inner.as_rule() {
-                    Rule::if_expr
-                    | Rule::for_expr
-                    | Rule::read_resource_expr
-                    | Rule::upstream_state_expr => Some(primary_inner.as_rule()),
-                    _ => None,
-                }
-            }
-            _ => None,
-        }
-    }
-    find_inner_rule(pair).is_some()
-}
-
-/// Parse expression with potential resource, module call, or import.
-fn parse_expression_with_resource_or_module(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-    binding_name: &str,
-) -> Result<LetBindingRhs, ParseError> {
-    let coalesce = first_inner(pair, "expression", "expression with resource or module")?;
-    let pipe = first_inner(coalesce, "pipe expression", "coalesce expression")?;
-    parse_pipe_expr_with_resource_or_module(pipe, ctx, binding_name)
-}
-
-fn parse_pipe_expr_with_resource_or_module(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-    binding_name: &str,
-) -> Result<LetBindingRhs, ParseError> {
-    let mut inner = pair.into_inner();
-    let compose_pair = next_pair(&mut inner, "compose expression", "pipe expression")?;
-
-    // Unwrap compose_expr: get its inner pairs
-    let mut compose_inner = compose_pair.into_inner();
-    let primary = next_pair(
-        &mut compose_inner,
-        "primary expression",
-        "compose expression",
-    )?;
-    let (mut value, expanded_resources, module_calls, maybe_import) =
-        parse_primary_with_resource_or_module(primary, ctx, binding_name)?;
-
-    // Handle >> composition within the compose_expr
-    let compose_rhs: Vec<_> = compose_inner.collect();
-    if !compose_rhs.is_empty() {
-        // Process the compose chain
-        for rhs_pair in compose_rhs {
-            let rhs = parse_primary_eval(rhs_pair, ctx)?;
-
-            if !value.is_closure() {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "left side of >> must be a Closure (partially applied function), got {}",
-                        eval_type_name(&value)
-                    ),
-                });
-            }
-            if !rhs.is_closure() {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "right side of >> must be a Closure (partially applied function), got {}",
-                        eval_type_name(&rhs)
-                    ),
-                });
-            }
-
-            let functions = if let EvalValue::Closure {
-                name,
-                captured_args,
-                ..
-            } = &value
-                && name == "__compose__"
-            {
-                let mut fns = captured_args.clone();
-                fns.push(rhs);
-                fns
-            } else {
-                vec![value, rhs]
-            };
-
-            value = EvalValue::closure("__compose__", functions, 1);
-        }
-    }
-
-    // Desugar pipe: `x |> f(args)` becomes `f(x, args)`
-    for func_call_pair in inner {
-        let mut fc_inner = func_call_pair.into_inner();
-        let func_name = next_pair(&mut fc_inner, "function name", "pipe function call")?
-            .as_str()
-            .to_string();
-        let extra_args: Result<Vec<Value>, ParseError> =
-            fc_inner.map(|arg| parse_expression(arg, ctx)).collect();
-        let extra_args = extra_args?;
-
-        // Lower the running pipe value to a `Value` for the builtin
-        // dispatch path (which expects fully-reduced data arguments).
-        // Closures are handled separately just below.
-        let pipe_value_for_args = match &value {
-            EvalValue::User(v) => Some(v.clone()),
-            EvalValue::Closure { .. } => None,
-        };
-
-        // Check if the pipe target is a Closure variable
-        if let Some(EvalValue::Closure {
-            name: fn_name,
-            captured_args,
-            remaining_arity,
-        }) = ctx.get_variable(&func_name)
-        {
-            // Build closure-application args. The pipe value (`x` in
-            // `x |> f`) goes as the last argument; we keep it as
-            // EvalValue so a chained closure can pipe through.
-            let mut all_args: Vec<EvalValue> = extra_args
-                .iter()
-                .cloned()
-                .map(EvalValue::from_value)
-                .collect();
-            all_args.push(value.clone());
-            if extra_args.iter().all(is_static_value)
-                && pipe_value_for_args.as_ref().is_some_and(is_static_value)
-            {
-                value = crate::builtins::apply_closure_with_config(
-                    fn_name,
-                    captured_args,
-                    *remaining_arity,
-                    &all_args,
-                    ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: e,
-                })?;
-                continue;
-            }
-        }
-
-        // Build args for the non-closure dispatch path: at this point
-        // we need a `Vec<Value>`, so the running pipe value must be a
-        // user-facing value (not a closure).
-        let pipe_value = match pipe_value_for_args {
-            Some(v) => v,
-            None => {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "cannot pipe a closure into '{}' — finish the partial application first",
-                        func_name
-                    ),
-                });
-            }
-        };
-        let mut args = extra_args;
-        args.push(pipe_value);
-
-        // Eagerly evaluate partial application for builtin pipe targets
-        if let Some(arity) = crate::builtins::builtin_arity(&func_name)
-            && args.len() < arity
-            && args.iter().all(is_static_value)
-        {
-            let eval_args: Vec<EvalValue> =
-                args.iter().cloned().map(EvalValue::from_value).collect();
-            value =
-                crate::builtins::evaluate_builtin_with_config(&func_name, &eval_args, ctx.config)
-                    .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("{}(): {}", func_name, e),
-                })?;
-            continue;
-        }
-
-        value = EvalValue::from_value(Value::FunctionCall {
-            name: func_name,
-            args,
-        });
-    }
-
-    Ok((value, expanded_resources, module_calls, maybe_import))
-}
-
-fn parse_primary_with_resource_or_module(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &mut ParseContext,
-    binding_name: &str,
-) -> Result<LetBindingRhs, ParseError> {
-    let inner = first_inner(pair, "value", "primary expression")?;
-
-    match inner.as_rule() {
-        Rule::read_resource_expr => {
-            let resource = parse_read_resource_expr(inner, ctx, binding_name)?;
-            let ref_value = Value::String(format!("${{{}}}", binding_name));
-            Ok((
-                EvalValue::from_value(ref_value),
-                vec![resource],
-                vec![],
-                None,
-            ))
-        }
-        Rule::upstream_state_expr => {
-            let (line, _) = inner.as_span().start_pos().line_col();
-            let us = parse_upstream_state_expr(inner, binding_name)?;
-            if ctx.upstream_states.contains_key(&us.binding) {
-                return Err(ParseError::DuplicateBinding {
-                    name: us.binding,
-                    line,
-                });
-            }
-            ctx.upstream_states.insert(us.binding.clone(), us);
-            let ref_value = Value::String(format!("${{{}}}", binding_name));
-            Ok((EvalValue::from_value(ref_value), vec![], vec![], None))
-        }
-        Rule::resource_expr => {
-            let resource = parse_resource_expr(inner, ctx, binding_name)?;
-            let ref_value = Value::String(format!("${{{}}}", binding_name));
-            Ok((
-                EvalValue::from_value(ref_value),
-                vec![resource],
-                vec![],
-                None,
-            ))
-        }
-        Rule::for_expr => {
-            let (resources, module_calls) = parse_for_expr(inner, ctx, binding_name)?;
-            let ref_value = Value::String(format!("${{for:{}}}", binding_name));
-            Ok((
-                EvalValue::from_value(ref_value),
-                resources,
-                module_calls,
-                None,
-            ))
-        }
-        Rule::if_expr => {
-            let (value, resources, module_calls, import) = parse_if_expr(inner, ctx, binding_name)?;
-            Ok((value, resources, module_calls, import))
-        }
-        Rule::module_call => {
-            let call = parse_module_call(inner, ctx)?;
-            let value = Value::String(format!("${{module:{}}}", call.module_name));
-            Ok((EvalValue::from_value(value), vec![], vec![call], None))
-        }
-        Rule::function_call => {
-            let value = parse_primary_eval(inner, ctx)?;
-            Ok((value, vec![], vec![], None))
-        }
-        _ => {
-            let value = parse_primary_eval(inner, ctx)?;
-            Ok((value, vec![], vec![], None))
-        }
-    }
-}
-
-/// Check whether a Value is fully static (no runtime dependencies).
-fn is_static_value(value: &Value) -> bool {
-    match value {
-        Value::String(_) | Value::Int(_) | Value::Float(_) | Value::Bool(_) => true,
-        Value::List(items) => items.iter().all(is_static_value),
-        Value::Map(map) => map.values().all(is_static_value),
-        Value::FunctionCall { args, .. } => args.iter().all(is_static_value),
-        Value::ResourceRef { .. } | Value::Interpolation(_) => false,
-        Value::Secret(inner) => is_static_value(inner),
-    }
-}
-
-/// `is_static_value` for the evaluator-internal `EvalValue` type.
-/// A closure's static-ness is decided by whether all of its captured
-/// args are themselves static. The pipe/compose paths use this when
-/// they need to decide whether to eagerly apply a partial application.
-fn is_static_eval(value: &EvalValue) -> bool {
-    match value {
-        EvalValue::User(v) => is_static_value(v),
-        EvalValue::Closure { captured_args, .. } => captured_args.iter().all(is_static_eval),
-    }
-}
-
-/// If `value` is a FunctionCall with all static arguments, eagerly evaluate it.
-/// Nested FunctionCalls in arguments are evaluated recursively first.
-fn evaluate_static_value(value: Value, config: &ProviderContext) -> Result<Value, ParseError> {
-    match value {
-        Value::FunctionCall { ref name, ref args } => {
-            if !is_static_value(&value) {
-                return Err(ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "for iterable function call '{name}' depends on a runtime value; \
-                         all arguments must be statically known at parse time"
-                    ),
-                });
-            }
-            // Recursively evaluate any nested FunctionCall arguments
-            let evaluated_args: Result<Vec<Value>, ParseError> = args
-                .iter()
-                .cloned()
-                .map(|v| evaluate_static_value(v, config))
-                .collect();
-            let evaluated_args = evaluated_args?;
-            let eval_args: Vec<EvalValue> = evaluated_args
-                .iter()
-                .cloned()
-                .map(EvalValue::from_value)
-                .collect();
-            let result = crate::builtins::evaluate_builtin_with_config(name, &eval_args, config)
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!("for iterable function call '{name}' failed: {e}"),
-                })?;
-            result
-                .into_value()
-                .map_err(|leak| ParseError::InvalidExpression {
-                    line: 0,
-                    message: format!(
-                        "for iterable function call '{name}' returned a closure '{}' \
-                     (still needs {} arg(s)); finish the partial application",
-                        leak.name, leak.remaining_arity
-                    ),
-                })
-        }
-        other => Ok(other),
-    }
-}
-
-fn parse_provider_block(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<ProviderConfig, ParseError> {
-    let mut inner = pair.into_inner();
-    let name = next_pair(&mut inner, "provider name", "provider block")?
-        .as_str()
-        .to_string();
-
-    let mut attributes: IndexMap<String, Value> = IndexMap::new();
-    for attr_pair in inner {
-        if attr_pair.as_rule() == Rule::attribute {
-            let mut attr_inner = attr_pair.into_inner();
-            let key = next_pair(&mut attr_inner, "attribute name", "provider block")?
-                .as_str()
-                .to_string();
-            let value = parse_expression(
-                next_pair(&mut attr_inner, "attribute value", "provider block")?,
-                ctx,
-            )?;
-            attributes.insert(key, value);
-        }
-    }
-
-    // `shift_remove` keeps the surviving attributes in source order.
-    // Extract default_tags from attributes if present
-    let default_tags = if let Some(Value::Map(tags)) = attributes.shift_remove("default_tags") {
-        tags
-    } else {
-        IndexMap::new()
-    };
-
-    // Extract source from attributes if present
-    let source = if let Some(Value::String(s)) = attributes.shift_remove("source") {
-        Some(s)
-    } else {
-        None
-    };
-
-    // Extract version from attributes if present
-    let version = if let Some(Value::String(v)) = attributes.shift_remove("version") {
-        Some(VersionConstraint::parse(&v).map_err(|e| {
-            pest::error::Error::new_from_pos(
-                pest::error::ErrorVariant::CustomError { message: e },
-                pest::Position::from_start(""),
-            )
-        })?)
-    } else {
-        None
-    };
-
-    // Extract revision from attributes if present
-    let revision = if let Some(Value::String(r)) = attributes.shift_remove("revision") {
-        Some(r)
-    } else {
-        None
-    };
-
-    // Validate that version and revision are mutually exclusive
-    if version.is_some() && revision.is_some() {
-        return Err(ParseError::Syntax(pest::error::Error::new_from_pos(
-            pest::error::ErrorVariant::CustomError {
-                message: format!(
-                    "Provider '{}': 'version' and 'revision' are mutually exclusive",
-                    name
-                ),
-            },
-            pest::Position::from_start(""),
-        )));
-    }
-
-    Ok(ProviderConfig {
-        name,
-        attributes,
-        default_tags,
-        source,
-        version,
-        revision,
-    })
-}
-
-/// Split a namespaced identifier (e.g., "awscc.ec2.Vpc") into (provider, resource_type)
-fn split_namespaced_id(namespaced: &str) -> (String, String) {
-    let parts: Vec<&str> = namespaced.split('.').collect();
-    if parts.len() >= 2 {
-        (parts[0].to_string(), parts[1..].join("."))
-    } else {
-        (String::new(), namespaced.to_string())
-    }
-}
-
-/// Parse a resource address: `provider.service.type "name"`
-fn parse_resource_address(pair: pest::iterators::Pair<Rule>) -> Result<ResourceId, ParseError> {
-    let mut inner = pair.into_inner();
-    let namespaced = next_pair(&mut inner, "namespaced id", "resource address")?
-        .as_str()
-        .to_string();
-    let name_pair = next_pair(&mut inner, "resource name", "resource address")?;
-    // The name is a string literal - extract value from quotes
-    let name = parse_string_literal(name_pair)?;
-
-    // Split namespaced id into provider and resource_type
-    let (provider, resource_type) = split_namespaced_id(&namespaced);
-
-    Ok(ResourceId::with_provider(provider, resource_type, name))
-}
-
-/// Parse an import state block
-fn parse_import_state_block(pair: pest::iterators::Pair<Rule>) -> Result<StateBlock, ParseError> {
-    let mut to: Option<ResourceId> = None;
-    let mut id: Option<String> = None;
-
-    for attr in pair.into_inner() {
-        if attr.as_rule() == Rule::import_state_attr {
-            let inner = first_inner(attr, "import attribute", "import block")?;
-            match inner.as_rule() {
-                Rule::import_to_attr => {
-                    let addr = first_inner(inner, "resource address", "import to")?;
-                    to = Some(parse_resource_address(addr)?);
-                }
-                Rule::import_id_attr => {
-                    let str_pair = first_inner(inner, "string", "import id")?;
-                    id = Some(parse_string_literal(str_pair)?);
-                }
-                _ => {}
-            }
-        }
-    }
-
-    let to = to.ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "import block requires 'to' attribute".to_string(),
-    })?;
-    let id = id.ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "import block requires 'id' attribute".to_string(),
-    })?;
-
-    Ok(StateBlock::Import { to, id })
-}
-
-/// Parse a removed block
-fn parse_removed_block(pair: pest::iterators::Pair<Rule>) -> Result<StateBlock, ParseError> {
-    let mut from: Option<ResourceId> = None;
-
-    for attr in pair.into_inner() {
-        if attr.as_rule() == Rule::removed_attr {
-            let addr = first_inner(attr, "resource address", "removed from")?;
-            from = Some(parse_resource_address(addr)?);
-        }
-    }
-
-    let from = from.ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "removed block requires 'from' attribute".to_string(),
-    })?;
-
-    Ok(StateBlock::Removed { from })
-}
-
-/// Parse a moved block
-fn parse_moved_block(pair: pest::iterators::Pair<Rule>) -> Result<StateBlock, ParseError> {
-    let mut from: Option<ResourceId> = None;
-    let mut to: Option<ResourceId> = None;
-
-    for attr in pair.into_inner() {
-        if attr.as_rule() == Rule::moved_attr {
-            let inner = first_inner(attr, "moved attribute", "moved block")?;
-            match inner.as_rule() {
-                Rule::moved_from_attr => {
-                    let addr = first_inner(inner, "resource address", "moved from")?;
-                    from = Some(parse_resource_address(addr)?);
-                }
-                Rule::moved_to_attr => {
-                    let addr = first_inner(inner, "resource address", "moved to")?;
-                    to = Some(parse_resource_address(addr)?);
-                }
-                _ => {}
-            }
-        }
-    }
-
-    let from = from.ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "moved block requires 'from' attribute".to_string(),
-    })?;
-    let to = to.ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "moved block requires 'to' attribute".to_string(),
-    })?;
-
-    Ok(StateBlock::Moved { from, to })
-}
-
-/// Parse a user-defined function definition
-fn parse_fn_def(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    warnings: &mut Vec<ParseWarning>,
-) -> Result<UserFunction, ParseError> {
-    let mut inner = pair.into_inner();
-    let name = next_pair(&mut inner, "function name", "fn_def")?
-        .as_str()
-        .to_string();
-
-    // Parse parameters (optional)
-    let mut params = Vec::new();
-    let next = next_pair(&mut inner, "fn_params or fn_body", "fn_def")?;
-    let next_token = if next.as_rule() == Rule::fn_params {
-        // Parse parameter list
-        for param_pair in next.into_inner() {
-            if param_pair.as_rule() == Rule::fn_param {
-                let mut param_inner = param_pair.into_inner();
-                let param_name = next_pair(&mut param_inner, "parameter name", "fn_param")?
-                    .as_str()
-                    .to_string();
-                // Parse optional type annotation (: type_expr)
-                let mut param_type = None;
-                let mut default = None;
-                for remaining in param_inner {
-                    match remaining.as_rule() {
-                        Rule::type_expr => {
-                            param_type = Some(parse_type_expr(remaining, ctx.config, warnings)?);
-                        }
-                        _ => {
-                            // This is the default expression
-                            let default_ctx = ParseContext::new(ctx.config);
-                            default = Some(parse_expression(remaining, &default_ctx)?);
-                        }
-                    }
-                }
-                // Validate: required params must come before optional params
-                if default.is_none() && params.iter().any(|p: &FnParam| p.default.is_some()) {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "in function '{name}': required parameter '{param_name}' cannot follow optional parameter"
-                    )));
-                }
-                params.push(FnParam {
-                    name: param_name,
-                    param_type,
-                    default,
-                });
-            }
-        }
-        next_pair(&mut inner, "type_expr or fn_body", "fn_def")?
-    } else {
-        next
-    };
-
-    // Parse optional return type annotation (: type_expr)
-    let (return_type, body_pair) = if next_token.as_rule() == Rule::type_expr {
-        let rt = parse_type_expr(next_token, ctx.config, warnings)?;
-        let bp = next_pair(&mut inner, "fn_body", "fn_def")?;
-        (Some(rt), bp)
-    } else {
-        (None, next_token)
-    };
-
-    // Parse body: fn_local_let* ~ (resource_expr | read_resource_expr | expression)
-    let mut local_lets = Vec::new();
-    let mut body: Option<UserFunctionBody> = None;
-
-    // Create a context where parameters are registered as variables
-    // so that param references in the body are resolved as variable refs
-    let mut body_ctx = ParseContext::new(ctx.config);
-    for p in &params {
-        body_ctx.set_variable(
-            p.name.clone(),
-            Value::String(format!("__fn_param_{}", p.name)),
-        );
-    }
-
-    for body_inner in body_pair.into_inner() {
-        match body_inner.as_rule() {
-            Rule::fn_local_let => {
-                let mut let_inner = body_inner.into_inner();
-                let let_name = next_pair(&mut let_inner, "let name", "fn_local_let")?
-                    .as_str()
-                    .to_string();
-                let let_expr = parse_expression(
-                    next_pair(&mut let_inner, "let expression", "fn_local_let")?,
-                    &body_ctx,
-                )?;
-                body_ctx.set_variable(
-                    let_name.clone(),
-                    Value::String(format!("__fn_local_{let_name}")),
-                );
-                local_lets.push((let_name, let_expr));
-            }
-            _ => {
-                // This should be the expression (the body)
-                body = Some(UserFunctionBody(parse_expression(body_inner, &body_ctx)?));
-            }
-        }
-    }
-
-    let body = body.ok_or_else(|| ParseError::InternalError {
-        expected: "body expression".to_string(),
-        context: "fn_def".to_string(),
-    })?;
-
-    Ok(UserFunction {
-        name,
-        params,
-        return_type,
-        local_lets,
-        body,
-    })
-}
-
-/// Prepare a user-defined function call: validate args, build substitutions, and return
-/// the child context with all parameters and local lets resolved.
-fn prepare_user_function_call<'cfg>(
-    func: &UserFunction,
-    args: &[Value],
-    ctx: &ParseContext<'cfg>,
-) -> Result<(ParseContext<'cfg>, HashMap<String, Value>), ParseError> {
-    let fn_name = &func.name;
-
-    // Check recursion
-    if ctx.evaluating_functions.contains(fn_name) {
-        return Err(ParseError::RecursiveFunction(fn_name.clone()));
-    }
-
-    // Validate argument count
-    let required_count = func.params.iter().filter(|p| p.default.is_none()).count();
-    let max_count = func.params.len();
-    if args.len() < required_count {
-        return Err(ParseError::UserFunctionError(format!(
-            "function '{fn_name}' expects at least {required_count} argument(s), got {}",
-            args.len()
-        )));
-    }
-    if args.len() > max_count {
-        return Err(ParseError::UserFunctionError(format!(
-            "function '{fn_name}' expects at most {max_count} argument(s), got {}",
-            args.len()
-        )));
-    }
-
-    // Build substitution map: param_name -> value, and type-check annotated params
-    let mut substitutions: HashMap<String, Value> = HashMap::new();
-    for (i, param) in func.params.iter().enumerate() {
-        let value = if i < args.len() {
-            args[i].clone()
-        } else {
-            param.default.clone().unwrap()
-        };
-        // Type-check if the parameter has a type annotation
-        if let Some(ref type_expr) = param.param_type {
-            check_fn_arg_type(fn_name, &param.name, type_expr, &value, ctx)?;
-        }
-        substitutions.insert(param.name.clone(), value);
-    }
-
-    // Create a child context with recursion tracking
-    let mut child_ctx = ctx.clone();
-    child_ctx.evaluating_functions.push(fn_name.clone());
-
-    // Evaluate local lets, substituting and resolving each one
-    for (let_name, let_expr) in &func.local_lets {
-        let substituted = substitute_fn_params(let_expr, &substitutions);
-        let evaluated = try_evaluate_fn_value(substituted, &child_ctx)?;
-        child_ctx.set_variable(let_name.clone(), evaluated.clone());
-        substitutions.insert(let_name.clone(), evaluated);
-    }
-
-    Ok((child_ctx, substitutions))
-}
-
-/// Validate a value against a custom type (ipv4_cidr, ipv4_address, etc.).
-/// Returns Ok(()) if the value passes validation or cannot be validated statically
-/// (e.g., ResourceRef, FunctionCall, Interpolation are deferred).
-///
-/// Checks built-in validators first, then falls back to custom validators
-/// registered in the [`ProviderContext`].
-pub fn validate_custom_type(
-    type_name: &str,
-    value: &Value,
-    config: &ProviderContext,
-) -> Result<(), String> {
-    match (type_name, value) {
-        ("ipv4_cidr", Value::String(s)) => validate_ipv4_cidr(s),
-        ("ipv4_address", Value::String(s)) => validate_ipv4_address(s),
-        ("ipv6_cidr", Value::String(s)) => validate_ipv6_cidr(s),
-        ("ipv6_address", Value::String(s)) => validate_ipv6_address(s),
-        (_, Value::ResourceRef { .. }) => Ok(()), // will be resolved later
-        (_, Value::FunctionCall { .. }) => Ok(()), // will be resolved later
-        (_, Value::Interpolation(_)) => Ok(()),   // will be resolved later
-        (name, Value::String(s)) => {
-            // Check custom validators from config (schema-extracted)
-            if let Some(validator) = config.validators.get(name) {
-                validator(s)?;
-            }
-            // Fall back to factory-based validator (e.g., WASM providers)
-            if let Some(ref factory_validator) = config.custom_type_validator {
-                factory_validator(name, s)
-            } else {
-                Ok(())
-            }
-        }
-        (_, value) => Err(format!(
-            "expected {}, got {}",
-            type_name,
-            value_type_name(value)
-        )),
-    }
-}
-
-/// Check that a function argument matches the declared parameter type.
-fn check_fn_arg_type(
-    fn_name: &str,
-    param_name: &str,
-    type_expr: &TypeExpr,
-    value: &Value,
-    ctx: &ParseContext,
-) -> Result<(), ParseError> {
-    let type_matches = match type_expr {
-        TypeExpr::String => matches!(
-            value,
-            Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-        ),
-        TypeExpr::Int => matches!(value, Value::Int(_)),
-        TypeExpr::Float => matches!(value, Value::Float(_)),
-        TypeExpr::Bool => matches!(value, Value::Bool(_)),
-        TypeExpr::List(_) => matches!(value, Value::List(_)),
-        TypeExpr::Map(_) => matches!(value, Value::Map(_)),
-        // Simple types (cidr, ipv4_address, arn, etc.) are string subtypes at runtime
-        TypeExpr::Simple(name) => {
-            if !matches!(
-                value,
-                Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-            ) {
-                false
-            } else {
-                // Validate the actual value against the custom type
-                if let Err(e) = validate_custom_type(name, value, ctx.config) {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "function '{fn_name}': parameter '{param_name}' type '{name}' validation failed: {e}"
-                    )));
-                }
-                true
-            }
-        }
-        // Resource type refs: check that the argument is a binding of the correct resource type
-        TypeExpr::Ref(expected_path) => {
-            // The argument is passed as a ResourceRef-like string "${binding_name}"
-            // or as a direct ResourceRef. Check if it corresponds to a resource binding
-            // of the expected type.
-            if let Value::String(s) = value
-                && let Some(ref_name) = s.strip_prefix("${").and_then(|s| s.strip_suffix('}'))
-                && let Some(resource) = ctx.resource_bindings.get(ref_name)
-            {
-                let actual_provider = &resource.id.provider;
-                let actual_type = &resource.id.resource_type;
-                if actual_provider != &expected_path.provider
-                    || actual_type != &expected_path.resource_type
-                {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "function '{fn_name}': parameter '{param_name}' expects resource type '{expected_path}', got {actual_provider}.{actual_type}"
-                    )));
-                }
-            }
-            // If not found in bindings, skip validation (forward ref or dynamic)
-            true
-        }
-        // Schema types (awscc.ec2.VpcId, etc.) are string subtypes with provider validators
-        TypeExpr::SchemaType { type_name, .. } => {
-            if !matches!(
-                value,
-                Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-            ) {
-                false
-            } else {
-                // Convert PascalCase type_name to snake_case for validator lookup
-                let validator_key = pascal_to_snake(type_name);
-                if let Err(e) = validate_custom_type(&validator_key, value, ctx.config) {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "function '{fn_name}': parameter '{param_name}' type '{type_expr}' validation failed: {e}"
-                    )));
-                }
-                true
-            }
-        }
-        TypeExpr::Struct { .. } => matches!(value, Value::Map(_)),
-    };
-    if !type_matches {
-        let actual_type = value_type_name(value);
-        return Err(ParseError::UserFunctionError(format!(
-            "function '{fn_name}': parameter '{param_name}' expects type '{type_expr}', got {actual_type}"
-        )));
-    }
-    Ok(())
-}
-
-/// Check that a function's return value matches the declared return type.
-fn check_fn_return_type(
-    fn_name: &str,
-    type_expr: &TypeExpr,
-    value: &Value,
-    config: &ProviderContext,
-) -> Result<(), ParseError> {
-    let type_matches = match type_expr {
-        TypeExpr::String => matches!(
-            value,
-            Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-        ),
-        TypeExpr::Int => matches!(value, Value::Int(_)),
-        TypeExpr::Float => matches!(value, Value::Float(_)),
-        TypeExpr::Bool => matches!(value, Value::Bool(_)),
-        TypeExpr::List(_) => matches!(value, Value::List(_)),
-        TypeExpr::Map(_) => matches!(value, Value::Map(_)),
-        // Simple types (cidr, ipv4_address, arn, etc.) — validate the value
-        TypeExpr::Simple(name) => {
-            if !matches!(
-                value,
-                Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-            ) {
-                false
-            } else {
-                if let Err(e) = validate_custom_type(name, value, config) {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "function '{fn_name}': return type '{name}' validation failed: {e}"
-                    )));
-                }
-                true
-            }
-        }
-        // Resource type refs: not applicable for value functions
-        TypeExpr::Ref(_) => true,
-        // Schema types: validate returned value against the provider validator
-        TypeExpr::SchemaType { type_name, .. } => {
-            if !matches!(
-                value,
-                Value::String(_) | Value::Interpolation(_) | Value::ResourceRef { .. }
-            ) {
-                false
-            } else {
-                let validator_key = pascal_to_snake(type_name);
-                if let Err(e) = validate_custom_type(&validator_key, value, config) {
-                    return Err(ParseError::UserFunctionError(format!(
-                        "function '{fn_name}': return type '{type_name}' validation failed: {e}"
-                    )));
-                }
-                true
-            }
-        }
-        TypeExpr::Struct { .. } => matches!(value, Value::Map(_)),
-    };
-    if !type_matches {
-        let actual_type = value_type_name(value);
-        return Err(ParseError::UserFunctionError(format!(
-            "function '{fn_name}': return type '{type_expr}' does not match actual return value of type {actual_type}"
-        )));
-    }
-    Ok(())
-}
-
-/// Convert PascalCase to snake_case (e.g., "VpcId" → "vpc_id", "SubnetId" → "subnet_id")
-pub fn pascal_to_snake(s: &str) -> String {
-    let mut result = String::with_capacity(s.len() + 4);
-    for (i, c) in s.chars().enumerate() {
-        if c.is_uppercase() {
-            if i > 0 {
-                result.push('_');
-            }
-            result.push(c.to_lowercase().next().unwrap());
-        } else {
-            result.push(c);
-        }
-    }
-    result
-}
-
-/// Convert snake_case to PascalCase (e.g., "vpc_id" → "VpcId", "aws_account_id" → "AwsAccountId").
-///
-/// Acronyms are treated as regular words (`iam_policy_arn` → `IamPolicyArn`,
-/// `ipv4_cidr` → `Ipv4Cidr`) so that the result matches `semantic_name` values
-/// already produced by `pascal_to_snake` and is a round-trip inverse for them.
-pub fn snake_to_pascal(s: &str) -> String {
-    let mut result = String::with_capacity(s.len());
-    let mut capitalize_next = true;
-    for c in s.chars() {
-        if c == '_' {
-            capitalize_next = true;
-        } else if capitalize_next {
-            result.push(c.to_ascii_uppercase());
-            capitalize_next = false;
-        } else {
-            result.push(c);
-        }
-    }
-    result
-}
-
-/// Return a human-readable type name for a Value
-pub(crate) fn value_type_name(value: &Value) -> &'static str {
-    match value {
-        Value::String(_) => "string",
-        Value::Int(_) => "int",
-        Value::Float(_) => "float",
-        Value::Bool(_) => "bool",
-        Value::List(_) => "list",
-        Value::Map(_) => "map",
-        Value::ResourceRef { .. } => "resource reference",
-        Value::Interpolation(_) => "string",
-        Value::FunctionCall { .. } => "function call",
-        Value::Secret(_) => "secret",
-    }
-}
-
-/// Return a human-readable type name for an `EvalValue`. Closures only
-/// exist on the evaluator-internal type, so this is the version used by
-/// pipe/compose error messages where a closure can legitimately show up.
-pub(crate) fn eval_type_name(value: &EvalValue) -> &'static str {
-    match value {
-        EvalValue::User(v) => value_type_name(v),
-        EvalValue::Closure { .. } => "closure",
-    }
-}
-
-/// Evaluate a user-defined function call by substituting arguments into the body
-fn evaluate_user_function(
-    func: &UserFunction,
-    args: &[Value],
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    let (child_ctx, substitutions) = prepare_user_function_call(func, args, ctx)?;
-
-    let UserFunctionBody(body) = &func.body;
-    let substituted_body = substitute_fn_params(body, &substitutions);
-    let result = try_evaluate_fn_value(substituted_body, &child_ctx)?;
-    // Check return type if annotated
-    if let Some(ref return_type) = func.return_type {
-        check_fn_return_type(&func.name, return_type, &result, child_ctx.config)?;
-    }
-    Ok(result)
-}
-
-/// Recursively substitute function parameter placeholders with actual values
-fn substitute_fn_params(value: &Value, substitutions: &HashMap<String, Value>) -> Value {
-    match value {
-        Value::String(s) => {
-            // Check if this is a parameter placeholder
-            if let Some(param_name) = s.strip_prefix("__fn_param_")
-                && let Some(sub) = substitutions.get(param_name)
-            {
-                return sub.clone();
-            }
-            if let Some(local_name) = s.strip_prefix("__fn_local_")
-                && let Some(sub) = substitutions.get(local_name)
-            {
-                return sub.clone();
-            }
-            Value::String(s.clone())
-        }
-        Value::List(items) => Value::List(
-            items
-                .iter()
-                .map(|v| substitute_fn_params(v, substitutions))
-                .collect(),
-        ),
-        Value::Map(map) => Value::Map(
-            map.iter()
-                .map(|(k, v)| (k.clone(), substitute_fn_params(v, substitutions)))
-                .collect(),
-        ),
-        Value::FunctionCall { name, args } => Value::FunctionCall {
-            name: name.clone(),
-            args: args
-                .iter()
-                .map(|a| substitute_fn_params(a, substitutions))
-                .collect(),
-        },
-        Value::Interpolation(parts) => Value::Interpolation(
-            parts
-                .iter()
-                .map(|p| match p {
-                    crate::resource::InterpolationPart::Expr(v) => {
-                        crate::resource::InterpolationPart::Expr(substitute_fn_params(
-                            v,
-                            substitutions,
-                        ))
-                    }
-                    other => other.clone(),
-                })
-                .collect(),
-        ),
-        Value::Secret(inner) => Value::Secret(Box::new(substitute_fn_params(inner, substitutions))),
-        other => other.clone(),
-    }
-}
-
-/// Try to evaluate a value (resolve function calls including user-defined ones)
-fn try_evaluate_fn_value(value: Value, ctx: &ParseContext) -> Result<Value, ParseError> {
-    match value {
-        Value::FunctionCall { ref name, ref args } => {
-            // First, recursively evaluate arguments
-            let evaluated_args: Result<Vec<Value>, ParseError> = args
-                .iter()
-                .map(|a| try_evaluate_fn_value(a.clone(), ctx))
-                .collect();
-            let evaluated_args = evaluated_args?;
-
-            // Check if the name refers to a Closure variable
-            if let Some(EvalValue::Closure {
-                name: fn_name,
-                captured_args,
-                remaining_arity,
-            }) = ctx.get_variable(name)
-            {
-                let eval_args: Vec<EvalValue> = evaluated_args
-                    .iter()
-                    .cloned()
-                    .map(EvalValue::from_value)
-                    .collect();
-                let result = crate::builtins::apply_closure_with_config(
-                    fn_name,
-                    captured_args,
-                    *remaining_arity,
-                    &eval_args,
-                    ctx.config,
-                )
-                .map_err(|e| ParseError::InvalidExpression {
-                    line: 0,
-                    message: e,
-                })?;
-                return result
-                    .into_value()
-                    .map_err(|leak| ParseError::InvalidExpression {
-                        line: 0,
-                        message: format!(
-                            "applying closure '{}' (still needs {} arg(s)) leaves a closure; \
-                         finish the partial application before using the result as data",
-                            leak.name, leak.remaining_arity
-                        ),
-                    });
-            }
-
-            // Try built-in first (with config for decrypt support)
-            let eval_args: Vec<EvalValue> = evaluated_args
-                .iter()
-                .cloned()
-                .map(EvalValue::from_value)
-                .collect();
-            match crate::builtins::evaluate_builtin_with_config(name, &eval_args, ctx.config) {
-                Ok(result) => result
-                    .into_value()
-                    .map_err(|leak| ParseError::InvalidExpression {
-                        line: 0,
-                        message: format!(
-                            "{}(): produced a closure '{}' (still needs {} arg(s)); \
-                         finish the partial application before using the result as data",
-                            name, leak.name, leak.remaining_arity
-                        ),
-                    }),
-                Err(_builtin_err) => {
-                    // Try user-defined function
-                    if let Some(user_fn) = ctx.user_functions.get(name) {
-                        evaluate_user_function(user_fn, &evaluated_args, ctx)
-                    } else {
-                        // Keep as FunctionCall (may contain unresolved refs)
-                        if evaluated_args.iter().all(is_static_value) {
-                            Err(ParseError::InvalidExpression {
-                                line: 0,
-                                message: format!("Unknown function: {name}"),
-                            })
-                        } else {
-                            Ok(Value::FunctionCall {
-                                name: name.clone(),
-                                args: evaluated_args,
-                            })
-                        }
-                    }
-                }
-            }
-        }
-        Value::List(items) => {
-            let evaluated: Result<Vec<Value>, ParseError> = items
-                .into_iter()
-                .map(|v| try_evaluate_fn_value(v, ctx))
-                .collect();
-            Ok(Value::List(evaluated?))
-        }
-        Value::Map(map) => {
-            let evaluated: Result<IndexMap<String, Value>, ParseError> = map
-                .into_iter()
-                .map(|(k, v)| try_evaluate_fn_value(v, ctx).map(|ev| (k, ev)))
-                .collect();
-            Ok(Value::Map(evaluated?))
-        }
-        Value::Interpolation(parts) => {
-            let evaluated: Result<Vec<crate::resource::InterpolationPart>, ParseError> = parts
-                .into_iter()
-                .map(|p| match p {
-                    crate::resource::InterpolationPart::Expr(v) => {
-                        try_evaluate_fn_value(v, ctx).map(crate::resource::InterpolationPart::Expr)
-                    }
-                    other => Ok(other),
-                })
-                .collect();
-            Ok(Value::Interpolation(evaluated?))
-        }
-        other => Ok(other),
-    }
-}
-
-fn parse_backend_block(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<BackendConfig, ParseError> {
-    let mut inner = pair.into_inner();
-    let backend_type = next_pair(&mut inner, "backend type", "backend block")?
-        .as_str()
-        .to_string();
-
-    let mut attributes = HashMap::new();
-    for attr_pair in inner {
-        if attr_pair.as_rule() == Rule::attribute {
-            let mut attr_inner = attr_pair.into_inner();
-            let key = next_pair(&mut attr_inner, "attribute name", "backend block")?
-                .as_str()
-                .to_string();
-            let value = parse_expression(
-                next_pair(&mut attr_inner, "attribute value", "backend block")?,
-                ctx,
-            )?;
-            attributes.insert(key, value);
-        }
-    }
-
-    Ok(BackendConfig {
-        backend_type,
-        attributes,
-    })
-}
-
-/// Parse an `upstream_state { source = "<dir>" }` expression.
-///
-/// The binding name comes from the enclosing `let` binding, so it's passed in
-/// rather than extracted from the expression itself.
-fn parse_upstream_state_expr(
-    pair: pest::iterators::Pair<Rule>,
-    binding_name: &str,
-) -> Result<UpstreamState, ParseError> {
-    let (block_line, _) = pair.as_span().start_pos().line_col();
-    let inner = pair.into_inner();
-
-    let mut source: Option<String> = None;
-    for attr_pair in inner {
-        if attr_pair.as_rule() != Rule::attribute {
-            continue;
-        }
-        let (attr_line, _) = attr_pair.as_span().start_pos().line_col();
-        let mut attr_inner = attr_pair.into_inner();
-        let key = next_pair(
-            &mut attr_inner,
-            "attribute name",
-            "upstream_state expression",
-        )?
-        .as_str()
-        .to_string();
-        let value_pair = next_pair(
-            &mut attr_inner,
-            "attribute value",
-            "upstream_state expression",
-        )?;
-        match key.as_str() {
-            "source" => {
-                let value_text = value_pair.as_str().to_string();
-                source = Some(extract_string_from_pair(value_pair).map_err(|_| {
-                    ParseError::InvalidExpression {
-                        line: attr_line,
-                        message: format!(
-                            "upstream_state '{}': 'source' must be a string literal, got: {}",
-                            binding_name, value_text
-                        ),
-                    }
-                })?);
-            }
-            other => {
-                return Err(ParseError::InvalidExpression {
-                    line: attr_line,
-                    message: format!(
-                        "unknown attribute '{}' in upstream_state '{}' expression",
-                        other, binding_name
-                    ),
-                });
-            }
-        }
-    }
-
-    let source = source.ok_or_else(|| ParseError::InvalidExpression {
-        line: block_line,
-        message: format!(
-            "upstream_state '{}' requires a 'source' attribute",
-            binding_name
-        ),
-    })?;
-
-    Ok(UpstreamState {
-        binding: binding_name.to_string(),
-        source: std::path::PathBuf::from(source),
-    })
-}
-
-/// Extract a string value from a pair (expression -> pipe_expr -> primary -> string)
-fn extract_string_from_pair(pair: pest::iterators::Pair<Rule>) -> Result<String, ParseError> {
-    // Walk through expression -> pipe_expr -> primary -> string -> string_part -> string_literal
-    fn find_string(pair: pest::iterators::Pair<Rule>) -> Option<String> {
-        if pair.as_rule() == Rule::string_literal {
-            return Some(pair.as_str().to_string());
-        }
-        if pair.as_rule() == Rule::single_quoted_content {
-            return Some(unescape_single_quoted(pair.as_str()));
-        }
-        if pair.as_rule() == Rule::single_quoted_string {
-            return pair
-                .into_inner()
-                .next()
-                .map(|p| unescape_single_quoted(p.as_str()));
-        }
-        if pair.as_rule() == Rule::string {
-            let mut result = String::new();
-            for inner in pair.into_inner() {
-                if let Some(s) = find_string(inner) {
-                    result.push_str(&s);
-                }
-            }
-            return Some(result);
-        }
-        if pair.as_rule() == Rule::string_part {
-            for inner in pair.into_inner() {
-                if let Some(s) = find_string(inner) {
-                    return Some(s);
-                }
-            }
-            return None;
-        }
-        for inner in pair.into_inner() {
-            if let Some(s) = find_string(inner) {
-                return Some(s);
-            }
-        }
-        None
-    }
-
-    find_string(pair).ok_or_else(|| ParseError::InvalidExpression {
-        line: 0,
-        message: "expected a string literal".to_string(),
-    })
-}
-
-fn parse_anonymous_resource(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Resource, ParseError> {
-    let inner = pair.into_inner();
-
-    let mut iter = inner;
-    let namespaced_type = next_pair(&mut iter, "resource type", "anonymous resource")?
-        .as_str()
-        .to_string();
-
-    // Extract resource type from namespace (aws.s3_bucket -> s3_bucket)
-    let parts: Vec<&str> = namespaced_type.split('.').collect();
-    if parts.len() < 2 {
-        return Err(ParseError::InvalidResourceType(namespaced_type));
-    }
-
-    let provider = parts[0];
-    let resource_type = parts[1..].join(".");
-
-    let mut quoted_out: Option<HashSet<String>> = Some(HashSet::new());
-    let attributes = parse_block_contents_with_quoted(iter, ctx, &mut quoted_out)?;
-    let quoted_string_attrs = quoted_out.unwrap_or_default();
-
-    // Anonymous resources get an empty name that will be replaced by a hash-based
-    // identifier computed from create-only properties after parsing.
-    let resource_name = String::new();
-
-    let mut attributes = attributes;
-    attributes.insert("_type".to_string(), Value::String(namespaced_type.clone()));
-
-    // Extract lifecycle block from attributes (it's a meta-argument, not a real attribute)
-    let lifecycle = extract_lifecycle_config(&mut attributes);
-
-    let id = ResourceId::with_provider(provider, resource_type, resource_name);
-
-    Ok(Resource {
-        id,
-        attributes: Expr::wrap_map(attributes),
-        kind: ResourceKind::Real,
-        lifecycle,
-        prefixes: HashMap::new(),
-        binding: None,
-        dependency_bindings: BTreeSet::new(),
-        module_source: None,
-        quoted_string_attrs,
-    })
-}
-
-/// True if a parsed expression is exactly a plain (uninterpolated)
-/// quoted string literal — no operators, no interpolation, no
-/// list / map wrapping. Used to populate `Resource.quoted_string_attrs`
-/// so enum-attribute diagnostics can distinguish a shape mismatch
-/// (`attr = "AWS_ACCOUNT"`) from a variant mismatch
-/// (`attr = AWS_ACCOUNT`); see #2094 / #2229.
-fn expression_is_plain_string_literal(pair: pest::iterators::Pair<Rule>) -> bool {
-    let Some(primary) = unwrap_to_primary(pair) else {
-        return false;
-    };
-    let Some(inner) = primary.into_inner().next() else {
-        return false;
-    };
-    if inner.as_rule() != Rule::string {
-        return false;
-    }
-    let Some(string_inner) = inner.into_inner().next() else {
-        return false;
-    };
-    match string_inner.as_rule() {
-        Rule::single_quoted_string => true,
-        Rule::double_quoted_string => string_inner
-            .into_inner()
-            .filter(|p| p.as_rule() == Rule::string_part)
-            .flat_map(|p| p.into_inner())
-            .all(|leaf| leaf.as_rule() != Rule::interpolation),
-        _ => false,
-    }
-}
-
-/// Walk an `expression` (or any intermediate operator rule) down to
-/// the first `primary`, returning `None` if any operator (`|>`, `>>`,
-/// `??`) has more than one operand — such chains are expressions, not
-/// pure literals.
-fn unwrap_to_primary(pair: pest::iterators::Pair<Rule>) -> Option<pest::iterators::Pair<Rule>> {
-    match pair.as_rule() {
-        Rule::primary => Some(pair),
-        Rule::expression | Rule::coalesce_expr | Rule::pipe_expr | Rule::compose_expr => {
-            let mut inner = pair.into_inner();
-            let first = inner.next()?;
-            if inner.next().is_some() {
-                return None;
-            }
-            unwrap_to_primary(first)
-        }
-        _ => None,
-    }
-}
-
-/// Parse block contents (attributes, nested blocks, and local let bindings)
-/// Nested blocks with the same name are collected into a list.
-/// Local let bindings are resolved within the block scope and NOT included in
-/// the returned attributes.
-fn parse_block_contents(
-    pairs: pest::iterators::Pairs<Rule>,
-    ctx: &ParseContext,
-) -> Result<IndexMap<String, Value>, ParseError> {
-    parse_block_contents_with_quoted(pairs, ctx, &mut None)
-}
-
-/// As [`parse_block_contents`], but if `quoted_out` is `Some`, populate it
-/// with the names of top-level attributes whose value is a plain quoted
-/// string literal (`attr = "..."`). Used by resource-level callers to
-/// build `Resource.quoted_string_attrs` for enum-attribute diagnostics
-/// (#2094 / #2229) without re-walking the pest tree.
-fn parse_block_contents_with_quoted(
-    pairs: pest::iterators::Pairs<Rule>,
-    ctx: &ParseContext,
-    quoted_out: &mut Option<HashSet<String>>,
-) -> Result<IndexMap<String, Value>, ParseError> {
-    // `IndexMap` so the order in which the user wrote attributes in the
-    // .crn file flows all the way to `Resource.attributes` and to
-    // `Value::Map` payloads — anything that re-renders attributes
-    // (formatter, plan display, diagnostics) sees a stable order.
-    let mut attributes: IndexMap<String, Value> = IndexMap::new();
-    let mut nested_blocks: IndexMap<String, Vec<Value>> = IndexMap::new();
-
-    // Local scope extends the parent context with block-scoped let bindings
-    let mut local_ctx = ctx.clone();
-
-    for content_pair in pairs {
-        match content_pair.as_rule() {
-            Rule::block_content => {
-                let inner = first_inner(content_pair, "block content item", "block content")?;
-                match inner.as_rule() {
-                    Rule::local_binding => {
-                        let mut binding_inner = inner.into_inner();
-                        let name =
-                            next_pair(&mut binding_inner, "binding name", "local let binding")?
-                                .as_str()
-                                .to_string();
-                        let value = parse_expression(
-                            next_pair(&mut binding_inner, "binding value", "local let binding")?,
-                            &local_ctx,
-                        )?;
-                        // Add to local scope only, not to attributes
-                        local_ctx.set_variable(name, value);
-                    }
-                    Rule::attribute => {
-                        let mut attr_inner = inner.into_inner();
-                        let key_pair =
-                            next_pair(&mut attr_inner, "attribute name", "block content")?;
-                        let key = extract_key_string(key_pair)?;
-                        let value_pair =
-                            next_pair(&mut attr_inner, "attribute value", "block content")?;
-                        record_quoted_if_literal(quoted_out, &key, &value_pair);
-                        let value = parse_expression(value_pair, &local_ctx)?;
-                        attributes.insert(key, value);
-                    }
-                    Rule::nested_block => {
-                        let mut block_inner = inner.into_inner();
-                        let block_name = next_pair(&mut block_inner, "block name", "nested block")?
-                            .as_str()
-                            .to_string();
-
-                        // Recursively parse nested block contents (supports arbitrary depth)
-                        let block_attrs = parse_block_contents(block_inner, &local_ctx)?;
-
-                        nested_blocks
-                            .entry(block_name)
-                            .or_default()
-                            .push(Value::Map(block_attrs));
-                    }
-                    _ => {}
-                }
-            }
-            Rule::attribute => {
-                let mut attr_inner = content_pair.into_inner();
-                let key_pair = next_pair(&mut attr_inner, "attribute name", "block content")?;
-                let key = extract_key_string(key_pair)?;
-                let value_pair = next_pair(&mut attr_inner, "attribute value", "block content")?;
-                record_quoted_if_literal(quoted_out, &key, &value_pair);
-                let value = parse_expression(value_pair, &local_ctx)?;
-                attributes.insert(key, value);
-            }
-            _ => {}
-        }
-    }
-
-    // Convert nested blocks to list attributes
-    for (name, blocks) in nested_blocks {
-        attributes.insert(name, Value::List(blocks));
-    }
-
-    Ok(attributes)
-}
-
-/// If `quoted_out` is enabled and `value_pair` is a plain quoted string
-/// literal (no interpolation, no operators, no list / map wrapping),
-/// record `key` in the output set.
-fn record_quoted_if_literal(
-    quoted_out: &mut Option<HashSet<String>>,
-    key: &str,
-    value_pair: &pest::iterators::Pair<Rule>,
-) {
-    if let Some(set) = quoted_out.as_mut()
-        && expression_is_plain_string_literal(value_pair.clone())
-    {
-        set.insert(key.to_string());
-    }
-}
-
-/// Extract lifecycle configuration from attributes.
-/// The parser parses `lifecycle { ... }` as a nested block, which becomes
-/// a List of Maps in attributes. We extract it and convert to LifecycleConfig.
-fn extract_lifecycle_config(attributes: &mut IndexMap<String, Value>) -> LifecycleConfig {
-    if let Some(Value::List(blocks)) = attributes.shift_remove("lifecycle") {
-        // Take the first lifecycle block (there should only be one)
-        if let Some(Value::Map(map)) = blocks.into_iter().next() {
-            let force_delete = matches!(map.get("force_delete"), Some(Value::Bool(true)));
-            let create_before_destroy =
-                matches!(map.get("create_before_destroy"), Some(Value::Bool(true)));
-            let prevent_destroy = matches!(map.get("prevent_destroy"), Some(Value::Bool(true)));
-            return LifecycleConfig {
-                force_delete,
-                create_before_destroy,
-                prevent_destroy,
-            };
-        }
-    }
-    LifecycleConfig::default()
-}
-
-fn parse_resource_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    binding_name: &str,
-) -> Result<Resource, ParseError> {
-    let mut inner = pair.into_inner();
-
-    let namespaced_type = next_pair(&mut inner, "resource type", "resource expression")?
-        .as_str()
-        .to_string();
-
-    // Extract resource type from namespace (aws.s3_bucket -> s3_bucket)
-    let parts: Vec<&str> = namespaced_type.split('.').collect();
-    if parts.len() < 2 {
-        return Err(ParseError::InvalidResourceType(namespaced_type));
-    }
-
-    // First part is provider name, the rest is resource type
-    let provider = parts[0];
-    let resource_type = parts[1..].join(".");
-
-    let mut quoted_out: Option<HashSet<String>> = Some(HashSet::new());
-    let mut attributes = parse_block_contents_with_quoted(inner, ctx, &mut quoted_out)?;
-    let quoted_string_attrs = quoted_out.unwrap_or_default();
-
-    // All providers: use binding name as identifier.
-    let resource_name = binding_name.to_string();
-
-    // Extract lifecycle block from attributes (it's a meta-argument, not a real attribute)
-    let lifecycle = extract_lifecycle_config(&mut attributes);
-
-    attributes.insert("_type".to_string(), Value::String(namespaced_type.clone()));
-
-    let id = ResourceId::with_provider(provider, resource_type, resource_name);
-
-    Ok(Resource {
-        id,
-        attributes: Expr::wrap_map(attributes),
-        kind: ResourceKind::Real,
-        lifecycle,
-        prefixes: HashMap::new(),
-        binding: Some(binding_name.to_string()),
-        dependency_bindings: BTreeSet::new(),
-        module_source: None,
-        quoted_string_attrs,
-    })
-}
-
-/// Parse a read resource expression (data source): read aws.s3_bucket { ... }
-fn parse_read_resource_expr(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-    binding_name: &str,
-) -> Result<Resource, ParseError> {
-    let mut inner = pair.into_inner();
-
-    let namespaced_type = next_pair(&mut inner, "resource type", "read resource expression")?
-        .as_str()
-        .to_string();
-
-    // Extract resource type from namespace (aws.s3_bucket -> s3_bucket)
-    let parts: Vec<&str> = namespaced_type.split('.').collect();
-    if parts.len() < 2 {
-        return Err(ParseError::InvalidResourceType(namespaced_type));
-    }
-
-    // First part is provider name, the rest is resource type
-    let provider = parts[0];
-    let resource_type = parts[1..].join(".");
-
-    let mut quoted_out: Option<HashSet<String>> = Some(HashSet::new());
-    let mut attributes = parse_block_contents_with_quoted(inner, ctx, &mut quoted_out)?;
-    let quoted_string_attrs = quoted_out.unwrap_or_default();
-
-    // All providers: use binding name as identifier.
-    let resource_name = binding_name.to_string();
-
-    // Extract lifecycle block from attributes (it's a meta-argument, not a real attribute)
-    let lifecycle = extract_lifecycle_config(&mut attributes);
-
-    attributes.insert("_type".to_string(), Value::String(namespaced_type.clone()));
-    // Mark as data source
-    attributes.insert("_data_source".to_string(), Value::Bool(true));
-
-    let id = ResourceId::with_provider(provider, resource_type, resource_name);
-
-    Ok(Resource {
-        id,
-        attributes: Expr::wrap_map(attributes),
-        kind: ResourceKind::DataSource,
-        lifecycle,
-        prefixes: HashMap::new(),
-        binding: Some(binding_name.to_string()),
-        dependency_bindings: BTreeSet::new(),
-        module_source: None,
-        quoted_string_attrs,
-    })
-}
-
-/// Parse an expression. The result is a fully-reduced `Value`: any
-/// closure that surfaces during evaluation surfaces here as a
-/// parse-time error. Use [`parse_expression_eval`] in pipe/compose
-/// paths where partial applications are legitimate intermediates.
-fn parse_expression(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<Value, ParseError> {
-    let eval = parse_expression_eval(pair, ctx)?;
-    eval.into_value()
-        .map_err(|leak| ParseError::InvalidExpression {
-            line: 0,
-            message: format!(
-                "expression evaluates to a closure '{}' (still needs {} arg(s)); finish the \
-             partial application — closures are not valid as data",
-                leak.name, leak.remaining_arity
-            ),
-        })
-}
-
-/// Parse an expression and return the raw `EvalValue`, preserving any
-/// closure produced during partial application. Only the pipe/compose
-/// paths and the let-binding RHS need this; everything else should
-/// call [`parse_expression`] and let unfinished closures surface as
-/// errors at the type boundary.
-fn parse_expression_eval(
-    pair: pest::iterators::Pair<Rule>,
-    ctx: &ParseContext,
-) -> Result<EvalValue, ParseError> {
-    let inner = first_inner(pair, "expression body", "expression")?;
-    parse_coalesce_expr(inner, ctx)
-}
-
-/// Build an `UndefinedIdentifier` error enriched with the best
-/// did-you-mean suggestion and the sorted in-scope binding list.
-fn undefined_identifier_error(
-    known: &std::collections::HashSet<&str>,
-    name: String,
-    line: usize,
-) -> ParseError {
-    ParseError::undefined_identifier(name, line, known.iter().map(|s| s.to_string()).collect())
-}
-
-impl ParseError {
-    /// Construct an `UndefinedIdentifier` error with did-you-mean suggestion
-    /// and sorted in-scope binding list. `known_bindings` is the set of
-    /// names that are in scope at the check site — they will be sorted
-    /// and used to compute the close-match suggestion. See #2038.
-    pub fn undefined_identifier(
-        name: String,
-        line: usize,
-        known_bindings: Vec<String>,
-    ) -> ParseError {
-        let known_refs: Vec<&str> = known_bindings.iter().map(String::as_str).collect();
-        let suggestion = crate::schema::suggest_similar_name(&name, &known_refs);
-        let mut in_scope = known_bindings;
-        in_scope.sort();
-        ParseError::UndefinedIdentifier {
-            name,
-            line,
-            suggestion,
-            in_scope,
-        }
-    }
-}
-
-/// Resolve forward references after the full binding set is known.
-///
-/// During single-pass parsing, `identifier.member` forms where `identifier` is
-/// not yet a known binding are stored as `String("identifier.member")`.
-/// This function walks all resource attributes, module call arguments, and attribute
-/// parameter values, converting matching strings to `ResourceRef`.
-fn resolve_forward_references(
-    resource_bindings: &HashMap<String, Resource>,
-    resources: &mut [Resource],
-    attribute_params: &mut [AttributeParameter],
-    module_calls: &mut [ModuleCall],
-    export_params: &mut [ExportParameter],
-) {
-    for resource in resources.iter_mut() {
-        // In-place replace via `iter_mut`: avoids the O(n²) cost of
-        // `shift_remove` + re-insert per key, and naturally preserves
-        // the user-authored attribute order without a key-collection
-        // round-trip. The placeholder is overwritten on the next line,
-        // so its identity doesn't matter.
-        for (_, expr) in resource.attributes.iter_mut() {
-            let placeholder = Value::Bool(false);
-            let value = std::mem::replace(&mut expr.0, placeholder);
-            expr.0 = resolve_forward_ref_in_value(value, resource_bindings);
-        }
-    }
-    for attr_param in attribute_params.iter_mut() {
-        if let Some(value) = attr_param.value.take() {
-            attr_param.value = Some(resolve_forward_ref_in_value(value, resource_bindings));
-        }
-    }
-    for call in module_calls.iter_mut() {
-        let keys: Vec<String> = call.arguments.keys().cloned().collect();
-        for key in keys {
-            if let Some(value) = call.arguments.remove(&key) {
-                let resolved = resolve_forward_ref_in_value(value, resource_bindings);
-                call.arguments.insert(key, resolved);
-            }
-        }
-    }
-    for export_param in export_params.iter_mut() {
-        if let Some(value) = export_param.value.take() {
-            export_param.value = Some(resolve_forward_ref_in_value(value, resource_bindings));
-        }
-    }
-}
-
-/// Recursively resolve forward references in a single Value.
-///
-/// Strings in `"name.member"` format where `name` is a known resource binding
-/// are resolved to `ResourceRef`. This handles forward references that were
-/// stored as strings during single-pass parsing.
-fn resolve_forward_ref_in_value(
-    value: Value,
-    resource_bindings: &HashMap<String, Resource>,
-) -> Value {
-    match value {
-        Value::String(ref s) => {
-            // A dotted string like "vpc.vpc_id" or "vpc.attr.nested" may be a
-            // forward reference that was stored as a string during single-pass
-            // parsing. Resolve it to ResourceRef if the first segment is a known
-            // resource binding. Parts after the second become field_path.
-            let parts: Vec<&str> = s.splitn(3, '.').collect();
-            if parts.len() >= 2 && resource_bindings.contains_key(parts[0]) {
-                let field_path = parts
-                    .get(2)
-                    .map(|rest| rest.split('.').map(|s| s.to_string()).collect())
-                    .unwrap_or_default();
-                return Value::resource_ref(parts[0].to_string(), parts[1].to_string(), field_path);
-            }
-            value
-        }
-        Value::List(items) => Value::List(
-            items
-                .into_iter()
-                .map(|v| resolve_forward_ref_in_value(v, resource_bindings))
-                .collect(),
-        ),
-        Value::Map(map) => Value::Map(
-            map.into_iter()
-                .map(|(k, v)| (k, resolve_forward_ref_in_value(v, resource_bindings)))
-                .collect(),
-        ),
-        Value::Interpolation(parts) => {
-            use crate::resource::InterpolationPart;
-            Value::Interpolation(
-                parts
-                    .into_iter()
-                    .map(|p| match p {
-                        InterpolationPart::Expr(v) => InterpolationPart::Expr(
-                            resolve_forward_ref_in_value(v, resource_bindings),
-                        ),
-                        other => other,
-                    })
-                    .collect(),
-            )
-        }
-        Value::FunctionCall { name, args } => Value::FunctionCall {
-            name,
-            args: args
-                .into_iter()
-                .map(|v| resolve_forward_ref_in_value(v, resource_bindings))
-                .collect(),
-        },
-        other => other,
-    }
-}
-
-/// Resolve resource references in a ParsedFile
-/// This replaces ResourceRef values with the actual attribute values from referenced resources
-pub fn resolve_resource_refs(parsed: &mut ParsedFile) -> Result<(), ParseError> {
-    resolve_resource_refs_with_config(parsed, &ProviderContext::default())
-}
-
-/// Resolve resource references with the given parser configuration.
-pub fn resolve_resource_refs_with_config(
-    parsed: &mut ParsedFile,
-    config: &ProviderContext,
-) -> Result<(), ParseError> {
-    // Save dependency bindings before resolution may change ResourceRef binding names.
-    // This preserves direct dependencies that would be lost by recursive resolution
-    // (e.g., tgw_attach.transit_gateway_id resolves to tgw.id, losing the tgw_attach dep).
-    for resource in &mut parsed.resources {
-        let deps = crate::deps::get_resource_dependencies(resource);
-        if !deps.is_empty() {
-            resource.dependency_bindings = deps.into_iter().collect();
-        }
-    }
-
-    // Build a map of binding_name -> attributes for quick lookup
-    let mut binding_map: HashMap<String, HashMap<String, Value>> = HashMap::new();
-    for resource in &parsed.resources {
-        if let Some(ref binding_name) = resource.binding {
-            // `binding_map` only needs key-based lookup, not source order
-            // (callers consume it via `.get(name)` for ResourceRef
-            // resolution), so the inner map stays `HashMap`.
-            binding_map.insert(binding_name.clone(), resource.resolved_attributes());
-        }
-    }
-
-    // Register argument parameters so they're recognized as valid bindings
-    for arg in &parsed.arguments {
-        binding_map.entry(arg.name.clone()).or_default();
-    }
-
-    // Register module call bindings so ResourceRefs to them are not rejected.
-    // The actual attribute values will be resolved after module expansion.
-    for call in &parsed.module_calls {
-        if let Some(ref name) = call.binding_name {
-            binding_map.entry(name.clone()).or_default();
-        }
-    }
-
-    // Register upstream_state bindings so ResourceRefs to them are not rejected.
-    // The actual attribute values will be resolved at plan time when the state file is loaded.
-    for us in &parsed.upstream_states {
-        binding_map.entry(us.binding.clone()).or_default();
-    }
-
-    // Resolve references in each resource. Keep `IndexMap` to preserve
-    // the user's source order through resolution (#2222).
-    for resource in &mut parsed.resources {
-        let mut resolved_attrs: IndexMap<String, Expr> = IndexMap::new();
-
-        for (key, expr) in &resource.attributes {
-            let resolved = resolve_value_with_config(expr, &binding_map, config)?;
-            resolved_attrs.insert(key.clone(), Expr(resolved));
-        }
-
-        resource.attributes = resolved_attrs;
-    }
-
-    // Resolve cross-file forward references in export_params.
-    // During per-file parsing, "binding.attribute" strings from sibling files
-    // remain as Value::String. Convert them to ResourceRef now that the full
-    // binding map is available.
-    let resource_bindings: HashMap<String, Resource> = parsed
-        .resources
-        .iter()
-        .filter_map(|r| r.binding.as_ref().map(|b| (b.clone(), r.clone())))
-        .collect();
-    for export_param in &mut parsed.export_params {
-        if let Some(value) = export_param.value.take() {
-            export_param.value = Some(resolve_forward_ref_in_value(value, &resource_bindings));
-        }
-    }
-
-    Ok(())
-}
-
-/// Every binding name declared in the merged `ParsedFile`: resources,
-/// arguments, module calls, upstream states, imports, user functions,
-/// variables, and for/if structural bindings.
-///
-/// This is the canonical answer to "is this identifier in scope?" for
-/// directory-wide checks. The same set feeds [`check_identifier_scope`]
-/// and the LSP borrows it (via `carina_lsp::diagnostics::checks`) to
-/// keep diagnostic suggestions consistent with the CLI.
-pub fn collect_known_bindings_merged(parsed: &ParsedFile) -> std::collections::HashSet<&str> {
-    let mut known: std::collections::HashSet<&str> = std::collections::HashSet::new();
-    known.extend(parsed.resources.iter().filter_map(|r| r.binding.as_deref())); // allow: direct — parser-internal, pre-expansion
-    known.extend(parsed.arguments.iter().map(|a| a.name.as_str()));
-    known.extend(
-        parsed
-            .module_calls
-            .iter()
-            .filter_map(|c| c.binding_name.as_deref()),
-    );
-    known.extend(parsed.upstream_states.iter().map(|u| u.binding.as_str()));
-    known.extend(parsed.uses.iter().map(|i| i.alias.as_str()));
-    known.extend(parsed.user_functions.keys().map(String::as_str));
-    known.extend(parsed.variables.keys().map(String::as_str));
-    known.extend(parsed.structural_bindings.iter().map(String::as_str));
-    known
-}
-
-/// Directory-wide identifier-scope validation for a merged [`ParsedFile`].
-///
-/// Emits one flat list of `UndefinedIdentifier` errors covering:
-///
-/// - Every `ResourceRef` whose root binding is not in scope (roots in
-///   resource attributes, attribute-parameter values, module-call
-///   arguments, and export-parameter values).
-/// - Every deferred for-expression iterable whose root is not in scope.
-///
-/// Errors are returned in a deterministic order: ResourceRef findings
-/// first (in resource / attribute / module / export order), then
-/// deferred-iterable findings. The caller (CLI `load_configuration_with_config`,
-/// LSP analysis pipeline) just inspects the returned `Vec` — both
-/// checks share the same `collect_known_bindings_merged` pass, so there
-/// is no performance reason to split them at the callsite.
-///
-/// **This is the canonical entry point for "is this identifier in
-/// scope?" checks.** Follow the #2104 rule: any new semantic check in
-/// that family gets added *here*, not as a new sibling function.
-pub fn check_identifier_scope(parsed: &ParsedFile) -> Vec<ParseError> {
-    let known = collect_known_bindings_merged(parsed);
-    let mut errors = Vec::new();
-    accumulate_undefined_reference_errors(parsed, &known, &mut errors);
-    accumulate_deferred_iterable_errors(parsed, &known, &mut errors);
-    errors
-}
-
-fn accumulate_undefined_reference_errors(
-    parsed: &ParsedFile,
-    known: &std::collections::HashSet<&str>,
-    errors: &mut Vec<ParseError>,
-) {
-    let mut seen: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-    let mut check = |value: &Value| {
-        value.visit_refs(&mut |path| {
-            let root = path.binding();
-            let root_ident = root.split(['[', ']']).next().unwrap_or(root);
-            if !known.contains(root_ident) && seen.insert(root_ident.to_string()) {
-                errors.push(undefined_identifier_error(known, root_ident.to_string(), 0));
-            }
-        });
-    };
-
-    for resource in &parsed.resources {
-        for expr in resource.attributes.values() {
-            check(&expr.0);
-        }
-    }
-    for attr in &parsed.attribute_params {
-        if let Some(value) = &attr.value {
-            check(value);
-        }
-    }
-    for call in &parsed.module_calls {
-        for v in call.arguments.values() {
-            check(v);
-        }
-    }
-    for export in &parsed.export_params {
-        if let Some(value) = &export.value {
-            check(value);
-        }
-    }
-}
-
-fn accumulate_deferred_iterable_errors(
-    parsed: &ParsedFile,
-    known: &std::collections::HashSet<&str>,
-    errors: &mut Vec<ParseError>,
-) {
-    for d in &parsed.deferred_for_expressions {
-        if !known.contains(d.iterable_binding.as_str()) {
-            errors.push(undefined_identifier_error(
-                known,
-                d.iterable_binding.clone(),
-                d.line,
-            ));
-        }
-    }
-}
-
-fn resolve_value_with_config(
-    value: &Value,
-    binding_map: &HashMap<String, HashMap<String, Value>>,
-    config: &ProviderContext,
-) -> Result<Value, ParseError> {
-    match value {
-        Value::ResourceRef { path } => match binding_map.get(path.binding()) {
-            Some(attributes) => match attributes.get(path.attribute()) {
-                Some(attr_value) => {
-                    // Recursively resolve in case the attribute itself is a reference
-                    resolve_value_with_config(attr_value, binding_map, config)
-                }
-                None => {
-                    // Attribute not found, keep as reference (might be resolved at runtime)
-                    Ok(value.clone())
-                }
-            },
-            None => Err(ParseError::UndefinedVariable(format!(
-                "{}.{}",
-                path.binding(),
-                path.attribute()
-            ))),
-        },
-        Value::List(items) => {
-            let resolved: Result<Vec<Value>, ParseError> = items
-                .iter()
-                .map(|item| resolve_value_with_config(item, binding_map, config))
-                .collect();
-            Ok(Value::List(resolved?))
-        }
-        Value::Map(map) => {
-            let mut resolved: IndexMap<String, Value> = IndexMap::new();
-            for (k, v) in map {
-                resolved.insert(
-                    k.clone(),
-                    resolve_value_with_config(v, binding_map, config)?,
-                );
-            }
-            Ok(Value::Map(resolved))
-        }
-        Value::Interpolation(parts) => {
-            use crate::resource::InterpolationPart;
-            let resolved: Result<Vec<InterpolationPart>, ParseError> = parts
-                .iter()
-                .map(|p| match p {
-                    InterpolationPart::Expr(v) => Ok(InterpolationPart::Expr(
-                        resolve_value_with_config(v, binding_map, config)?,
-                    )),
-                    other => Ok(other.clone()),
-                })
-                .collect();
-            Ok(Value::Interpolation(resolved?))
-        }
-        Value::FunctionCall { name, args } => {
-            let resolved_args: Result<Vec<Value>, ParseError> = args
-                .iter()
-                .map(|a| resolve_value_with_config(a, binding_map, config))
-                .collect();
-            let resolved_args = resolved_args?;
-
-            let all_args_resolved = resolved_args.iter().all(is_static_value);
-
-            let eval_args: Vec<EvalValue> = resolved_args
-                .iter()
-                .cloned()
-                .map(EvalValue::from_value)
-                .collect();
-            match crate::builtins::evaluate_builtin_with_config(name, &eval_args, config) {
-                Ok(result) => result
-                    .into_value()
-                    .map_err(|leak| ParseError::InvalidExpression {
-                        line: 0,
-                        message: format!(
-                            "{}(): produced a closure '{}' (still needs {} arg(s)); \
-                         finish the partial application before using the result as data",
-                            name, leak.name, leak.remaining_arity
-                        ),
-                    }),
-                Err(e) => {
-                    if all_args_resolved {
-                        // All args are resolved but builtin failed — propagate the error
-                        Err(ParseError::InvalidExpression {
-                            line: 0,
-                            message: format!("{}(): {}", name, e),
-                        })
-                    } else {
-                        // Args contain unresolved refs — keep as FunctionCall for later resolution
-                        Ok(Value::FunctionCall {
-                            name: name.clone(),
-                            args: resolved_args,
-                        })
-                    }
-                }
-            }
-        }
-        _ => Ok(value.clone()),
-    }
-}
-
-/// Parse a .crn file and resolve resource references
-pub fn parse_and_resolve(input: &str) -> Result<ParsedFile, ParseError> {
-    let mut parsed = parse(input, &ProviderContext::default())?;
-    resolve_resource_refs(&mut parsed)?;
-    Ok(parsed)
-}
+pub(super) struct CarinaParser;
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::resource::InterpolationPart;
+    use crate::resource::{InterpolationPart, Resource, Value};
+    use indexmap::IndexMap;
+    use std::collections::HashMap;
 
     #[test]
     fn parse_and_resolve_returns_value_only_no_closure() {
